@@ -21,69 +22,54 @@
     "pretty:html:fix": "prettier --no-config --print-width=140 --write \"{src,lib}/**/**.html\"",
     "clean": "rm -rf node_modules/ && npm install"
   },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged",
-      "pre-push": "ng lint --fix"
-    }
-  },
-  "lint-staged": {
-    "**/*.{js,ts,json,css,scss,html,md}": [
-      "prettier --write",
-      "git add"
-    ]
-  },
   "dependencies": {
-    "@angular-material-components/datetime-picker": "^6.0.3",
-    "@angular/animations": "~12.2.14",
-    "@angular/cdk": "~12.2.13",
-    "@angular/common": "~12.2.14",
-    "@angular/compiler": "~12.2.14",
-    "@angular/core": "~12.2.14",
-    "@angular/flex-layout": "^12.0.0-beta.35",
-    "@angular/forms": "~12.2.14",
-    "@angular/material": "~12.2.13",
-    "@angular/platform-browser": "~12.2.14",
-    "@angular/platform-browser-dynamic": "~12.2.14",
-    "@angular/router": "~12.2.14",
-    "@kolkov/angular-editor": "^1.2.0",
-    "@ngx-translate/core": "^13.0.0",
-    "@types/nouislider": "^9.0.7",
-    "lodash": "^4.17.20",
-    "moment": "^2.29.0",
-    "ngx-markdown": "^12.0.1",
-    "ngx-mask": "12.0.0",
-    "nouislider": "^14.7.0",
-    "rxjs": "~6.6.7",
+    "@angular/animations": "^17.3.5",
+    "@angular/cdk": "^17.3.5",
+    "@angular/common": "^17.3.5",
+    "@angular/compiler": "^17.3.5",
+    "@angular/core": "^17.3.5",
+    "@angular/forms": "^17.3.5",
+    "@angular/material": "^17.3.5",
+    "@angular/platform-browser": "^17.3.5",
+    "@angular/platform-browser-dynamic": "^17.3.5",
+    "@angular/router": "^17.3.5",
+    "@ngx-translate/core": "^15.0.0",
+    "ngx-markdown": "^17.2.0",
+    "prismjs": "^1.29.0",
+    "rxjs": "~7.8.1",
     "tslib": "^2.0.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "~0.14.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.2.13",
-    "@angular/cli": "~12.2.13",
-    "@angular/compiler-cli": "~12.2.14",
-    "@angular/language-service": "~12.2.14",
+    "@angular-devkit/build-angular": "^17.3.5",
+    "@angular-eslint/builder": "17.3.0",
+    "@angular-eslint/eslint-plugin": "17.3.0",
+    "@angular-eslint/eslint-plugin-template": "17.3.0",
+    "@angular-eslint/schematics": "17.3.0",
+    "@angular-eslint/template-parser": "17.3.0",
+    "@angular/cli": "^17.3.5",
+    "@angular/compiler-cli": "^17.3.5",
+    "@angular/language-service": "^17.3.5",
     "@types/jasmine": "~3.10.2",
     "@types/jasminewd2": "~2.0.10",
-    "@types/node": "^12.11.1",
-    "angular-cli-ghpages": "^1.0.0-rc.2",
-    "codelyzer": "^6.0.0",
-    "cz-conventional-changelog": "^3.3.0",
-    "husky": "^4.3.8",
+    "@types/node": "^20.11.1",
+    "@typescript-eslint/eslint-plugin": "^7.2.0",
+    "@typescript-eslint/parser": "^7.2.0",
+    "angular-cli-ghpages": "^2.0.0-beta.2",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.1.3",
     "jasmine-core": "~3.10.1",
     "jasmine-spec-reporter": "~7.0.0",
-    "karma": "~6.3.2",
+    "karma": "~6.4.1",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "lint-staged": "^10.5.4",
-    "ng-packagr": "^12.2.5",
-    "prettier": "2.3.1",
+    "ng-packagr": "^17.3.0",
+    "prettier": "3.2.5",
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
-    "tslint": "~6.1.0",
-    "tslint-plugin-prettier": "^2.3.0",
-    "typescript": "~4.3.5"
+    "typescript": "~5.4.5"
   }
 }
