@@ -1,10 +1,10 @@
-import { Injectable } from '@angular/core';
+import { inject, Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
 
 @Injectable({ providedIn: 'root' })
 export class FileService {
-  public constructor(private http: HttpClient) {}
+  private readonly http = inject(HttpClient);
 
   public fetchFile(path: string): Observable<string> {
     return this.http.get(path, { responseType: 'text' });
