@@ -1,28 +1,28 @@
-import { Component, Input } from '@angular/core';
+import {
+  ChangeDetectionStrategy,
+  Component,
+  Input,
+  ViewEncapsulation,
+} from '@angular/core';
 import { BreadCrumb } from '../../models/bread-crumb.model';
-import { readPropValue } from '../../../utils/utils';
+
+import { RouterLink } from '@angular/router';
+import { TranslateModule } from '@ngx-translate/core';
+import { BreadCrumbItemComponent } from '../bread-crumb-item/bread-crumb-item.component';
 
 @Component({
   selector: 'lab900-bread-crumbs',
   templateUrl: './bread-crumbs.component.html',
   styleUrls: ['./bread-crumbs.component.scss'],
+  standalone: true,
+  imports: [RouterLink, TranslateModule, BreadCrumbItemComponent],
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  encapsulation: ViewEncapsulation.None,
 })
 export class BreadCrumbsComponent {
-  @Input()
+  @Input({ required: true })
   public breadCrumbs: BreadCrumb[];
 
   @Input()
   public data?: any;
-
-  public getTitle(title: ((data: any) => string) | string): string | null {
-    return readPropValue(title, this.data);
-  }
-
-  public getRoute(route: ((data: any) => string) | string): string | null {
-    return readPropValue(route, this.data);
-  }
-
-  public getQueryParams(queryParams: ((data: any) => object) | object | undefined): object | null {
-    return readPropValue(queryParams, this.data);
-  }
 }
