@@ -11,7 +11,9 @@ export abstract class SubscriptionBasedDirective implements OnDestroy {
     errorCallback?: (error: any) => void,
     completeCallback?: () => void,
   ): void {
-    this.subscriptions.push(observable.subscribe(successCallback, errorCallback, completeCallback));
+    this.subscriptions.push(
+      observable.subscribe(successCallback, errorCallback, completeCallback),
+    );
   }
 
   public ngOnDestroy(): void {
