@@ -1,44 +1,74 @@
-import { Component, Input } from '@angular/core';
+import {
+  ChangeDetectionStrategy,
+  Component,
+  Input,
+  ViewEncapsulation,
+} from '@angular/core';
 import { NavItem } from '../../models/nav-item.model';
+import {
+  BehaviorSubject,
+  combineLatest,
+  Observable,
+  ReplaySubject,
+} from 'rxjs';
+import { map } from 'rxjs/operators';
+import { AsyncPipe, NgClass, NgTemplateOutlet } from '@angular/common';
+import { IconComponent } from '../icon/icon.component';
+import { TranslateModule } from '@ngx-translate/core';
+import { MatListModule } from '@angular/material/list';
+import {
+  IsActiveMatchOptions,
+  RouterLink,
+  RouterLinkActive,
+} from '@angular/router';
 
 @Component({
   selector: 'lab900-nav-item-button',
   styleUrls: ['./nav-item-button.component.scss'],
-  template: `<button
-    fxFlex="1 1 auto"
-    role="menuitem"
-    class="nav-item-btn"
-    mat-list-item
-    [disabled]="disabled"
-    [ngStyle]="{ 'padding-left': depth * 12 + 'px' }"
-    [ngClass]="{ active: active, expanded: expanded, expandable: item?.children?.length }"
-  >
-    <lab900-icon
-      class="nav-item-btn__icon-left"
-      [icon]="item.icon"
-      *ngIf="item.icon && (item.icon?.position === 'left' || !item.icon?.position)"
-    ></lab900-icon>
-    <span fxFlex="1 1 auto">{{ item.label | translate }}</span>
-    <lab900-icon
-      class="nav-item-btn__icon-right"
-      [icon]="item.icon"
-      *ngIf="!showLevelArrow && item?.icon?.position === 'right'"
-    ></lab900-icon>
-    <lab900-icon class="nav-item-btn__icon-right" [icon]="{ name: levelArrow }" *ngIf="showLevelArrow"></lab900-icon>
-  </button>`,
+  templateUrl: './nav-item-button.component.html',
+  encapsulation: ViewEncapsulation.None,
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: true,
+  imports: [
+    AsyncPipe,
+    IconComponent,
+    TranslateModule,
+    MatListModule,
+    NgClass,
+    NgTemplateOutlet,
+    RouterLink,
+    RouterLinkActive,
+  ],
 })
 export class NavItemButtonComponent {
-  @Input()
-  public readonly item: NavItem;
+  private readonly _item$ = new ReplaySubject<NavItem>();
+  public readonly item$: Observable<NavItem> = this._item$.asObservable();
+
+  @Input({ required: true })
+  public set item(item: NavItem) {
+    this._item$.next(item);
+  }
+
+  private readonly _disabled$ = new BehaviorSubject<boolean>(false);
+  public readonly disabled$: Observable<boolean> =
+    this._disabled$.asObservable();
 
   @Input()
-  public readonly expanded: boolean;
+  public set disabled(disabled: boolean) {
+    this._disabled$.next(disabled);
+  }
+
+  private readonly _expanded$ = new BehaviorSubject<boolean>(false);
+  public readonly expanded$: Observable<boolean> =
+    this._expanded$.asObservable();
 
   @Input()
-  public readonly active: boolean;
+  public set expanded(expanded: boolean) {
+    this._expanded$.next(expanded);
+  }
 
   @Input()
-  public readonly disabled: boolean;
+  public navListMatchOptions?: IsActiveMatchOptions | { exact: boolean };
 
   @Input()
   public readonly depth: number;
@@ -46,10 +76,15 @@ export class NavItemButtonComponent {
   @Input()
   public readonly showLevelArrow: boolean = false;
 
-  get levelArrow(): string {
-    if (this.item?.childrenInOverlay) {
-      return 'navigate_next';
-    }
-    return this.expanded ? 'expand_less' : 'expand_more';
+  public readonly icon$: Observable<string>;
+  public constructor() {
+    this.icon$ = combineLatest([this.item$, this.expanded$]).pipe(
+      map(([item, expanded]) => {
+        if (item?.childrenInOverlay) {
+          return 'navigate_next';
+        }
+        return expanded ? 'expand_less' : 'expand_more';
+      }),
+    );
   }
 }
