@@ -1,19 +1,33 @@
-import { Component, Input, OnDestroy, OnInit } from '@angular/core';
-import { Event, NavigationEnd, Router } from '@angular/router';
+import {
+  ChangeDetectionStrategy,
+  Component,
+  Input,
+  OnInit,
+  ViewEncapsulation,
+} from '@angular/core';
+import { IsActiveMatchOptions, NavigationEnd, Router } from '@angular/router';
 import { NavItem } from '../../models/nav-item.model';
-import { Subscription } from 'rxjs';
+import { BehaviorSubject, Observable } from 'rxjs';
 import { SubscriptionBasedDirective } from '../../../common/directives/subscription-based.directive';
-import { MediaObserver } from '@angular/flex-layout';
+import { filter } from 'rxjs/operators';
+import { BreakpointObserver, Breakpoints } from '@angular/cdk/layout';
+import { AsyncPipe } from '@angular/common';
+import { NavItemButtonComponent } from '../nav-item-button/nav-item-button.component';
 
 @Component({
   selector: 'lab900-nav-item',
   templateUrl: './nav-item.component.html',
   styleUrls: ['./nav-item.component.scss'],
+  encapsulation: ViewEncapsulation.None,
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: true,
+  imports: [NavItemButtonComponent, AsyncPipe],
 })
-export class NavItemComponent extends SubscriptionBasedDirective implements OnInit, OnDestroy {
-  private sub: Subscription;
-
-  @Input()
+export class NavItemComponent
+  extends SubscriptionBasedDirective
+  implements OnInit
+{
+  @Input({ required: true })
   public readonly item: NavItem;
 
   @Input()
@@ -26,51 +40,61 @@ export class NavItemComponent extends SubscriptionBasedDirective implements OnIn
   public readonly depth = 0;
 
   @Input()
-  public readonly allowOverlayMenuUntil: string | string[] = 'xs';
+  public readonly allowOverlayMenuUntil: string | string[] = Breakpoints.XSmall;
+
+  private readonly _disabled$ = new BehaviorSubject<boolean>(false);
+  public readonly disabled$: Observable<boolean> =
+    this._disabled$.asObservable();
 
   @Input()
-  public disabled = false;
+  public set disabled(disabled: boolean) {
+    this._disabled$.next(disabled);
+  }
+
+  private readonly _expanded$ = new BehaviorSubject<boolean>(false);
+  public readonly expanded$: Observable<boolean> =
+    this._expanded$.asObservable();
 
   @Input()
-  public expanded = false;
+  public set expanded(expanded: boolean) {
+    this._expanded$.next(expanded);
+  }
 
-  public constructor(public readonly router: Router, public readonly mediaObserver: MediaObserver) {
+  @Input()
+  public navListMatchOptions?: IsActiveMatchOptions | { exact: boolean };
+
+  public constructor(
+    public readonly router: Router,
+    public readonly breakpointObserver: BreakpointObserver,
+  ) {
     super();
   }
 
   public ngOnInit(): void {
     if (!(this.item.route || this.item.href || this.item.children)) {
-      this.disabled = true;
+      this._disabled$.next(true);
     } else {
-      this.addSubscription(this.router.events, (event: Event) => {
-        if (event instanceof NavigationEnd) {
+      this.addSubscription(
+        this.router.events.pipe(filter((e) => e instanceof NavigationEnd)),
+        (event: NavigationEnd) => {
+          this.item.navigationFinished?.(true);
           const url = event.urlAfterRedirects;
           if (url && this.item?.children?.length) {
-            this.expanded = this.item.children.some((item: NavItem) => url.indexOf(`/${item.route}`) === 0);
+            this._expanded$.next(
+              this.item.children.some(
+                (item: NavItem) => url.indexOf(`/${item.route}`) === 0,
+              ),
+            );
           }
-        }
-      });
-    }
-  }
-
-  public ngOnDestroy(): void {
-    if (this.sub) {
-      this.sub.unsubscribe();
+        },
+      );
     }
   }
 
   public onClick(event: MouseEvent): void {
-    event.preventDefault();
     if (this.item?.children?.length) {
-      this.expanded = !this.expanded;
-    } else if (this.item?.route) {
-      this.router.navigate([this.item.route], { queryParams: this.item?.routeQueryParams }).then((navigationResult: boolean) => {
-        if (this.item?.navigationFinished) {
-          this.item.navigationFinished(navigationResult);
-        }
-      });
-    } else if (this.item?.href?.url) {
-      window.open(this.item.href.url, this.item.href?.target ?? '_blank');
+      event.preventDefault();
+      this._expanded$.next(!this._expanded$.value);
     }
   }
 }
