@@ -1,50 +1,81 @@
-import { Component, Input, OnChanges, OnInit, SimpleChanges } from '@angular/core';
+import {
+  ChangeDetectionStrategy,
+  Component,
+  Input,
+  ViewEncapsulation,
+} from '@angular/core';
 import { PageHeaderNavItem } from '../../models/page-header-nav.model';
 import { ActionButton } from '../../../button/models/action-button.model';
 import { BreadCrumb } from '../../../bread-crumbs/models/bread-crumb.model';
-import { readPropValue } from '../../../utils/utils';
+import { MatTabNavPanel, MatTabsModule } from '@angular/material/tabs';
+import { Observable, ReplaySubject } from 'rxjs';
+import { map } from 'rxjs/operators';
+import { PageHeaderNavItemComponent } from '../page-header-nav-item/page-header-nav-item.component';
+import { AsyncPipe } from '@angular/common';
+import { MatListModule } from '@angular/material/list';
+import { TranslateModule } from '@ngx-translate/core';
+import { BreadCrumbsComponent } from '../../../bread-crumbs/components/bread-crumbs/bread-crumbs.component';
+import { Lab900ActionButtonComponent } from '../../../button/components/action-button/lab900-action-button.component';
 
 @Component({
   selector: 'lab900-page-header',
   templateUrl: './page-header.component.html',
   styleUrls: ['./page-header.component.scss'],
+  encapsulation: ViewEncapsulation.None,
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: true,
+  imports: [
+    MatTabsModule,
+    PageHeaderNavItemComponent,
+    AsyncPipe,
+    MatListModule,
+    BreadCrumbsComponent,
+    TranslateModule,
+    Lab900ActionButtonComponent,
+  ],
 })
-export class Lab900PageHeaderComponent implements OnChanges {
+export class Lab900PageHeaderComponent {
   @Input()
-  public pageTitle: string;
+  public pageTitle?: string;
 
   @Input()
   public pageTitleArgs: object = {};
 
   @Input()
-  public navItems: PageHeaderNavItem[];
+  public navItems?: PageHeaderNavItem[];
+
+  /**
+   * The corresponding <router-outlet> must be wrapped in an <mat-tab-nav-panel> component and should typically be placed relatively close to this component
+   */
+  @Input()
+  public tabPanel?: MatTabNavPanel;
 
   @Input()
-  public actions: ActionButton[];
-  public leftActions: ActionButton[];
-  public rightActions: ActionButton[];
+  public showActionsFixedOnMobile = true;
+
+  @Input()
+  public navFitInkBarToContent = false;
+  public navStretch = false;
+
+  private readonly _actions$ = new ReplaySubject<ActionButton[]>();
+  public readonly actions$: Observable<ActionButton[]> =
+    this._actions$.asObservable();
+  public readonly leftActions$ = this.actions$.pipe(
+    map(([...actions]) => actions?.filter((action) => action.align === 'left')),
+  );
+
+  public readonly rightActions$ = this.actions$.pipe(
+    map(([...actions]) => actions?.filter((action) => action.align !== 'left')),
+  );
+
+  @Input()
+  public set actions(actions: ActionButton[]) {
+    this._actions$.next(actions);
+  }
 
   @Input()
   public data?: any;
 
   @Input()
-  public breadCrumbs: BreadCrumb[];
-
-  public ngOnChanges(changes: SimpleChanges): void {
-    if (changes.actions) {
-      this.leftActions = this.actions?.filter((action) => action.align === 'left');
-      this.rightActions = this.actions?.filter((action) => action.align === 'right' || action.align == null);
-    }
-  }
-
-  public getLabel(item: PageHeaderNavItem): string {
-    if (typeof item.label === 'function') {
-      return item.label(this.data);
-    }
-    return item.label;
-  }
-
-  public getRoute(route: ((data: any) => string) | string): string | null {
-    return readPropValue(route, this.data);
-  }
+  public breadCrumbs?: BreadCrumb[];
 }
