@@ -1,7 +1,14 @@
 import { Component, Input } from '@angular/core';
-import { MatDialog } from '@angular/material/dialog';
-import { ShareDialogComponent, SharingDialogData } from '../../models/share-dialog-component.abstract';
+import { MatDialog, MatDialogModule } from '@angular/material/dialog';
+import {
+  ShareDialogComponent,
+  SharingDialogData,
+} from '../../models/share-dialog-component.abstract';
 import { ComponentType } from '@angular/cdk/portal';
+
+import { MatButtonModule } from '@angular/material/button';
+import { MatIconModule } from '@angular/material/icon';
+import { MatTooltipModule } from '@angular/material/tooltip';
 
 const materialIconBase64 =
   'data:image/svg+xml;utf-8,<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="black" width="18px" height="18px"><path d="M0 0h24v24H0z" fill="none"/><path d="M9 11.75c-.69 0-1.25.56-1.25 1.25s.56 1.25 1.25 1.25 1.25-.56 1.25-1.25-.56-1.25-1.25-1.25zm6 0c-.69 0-1.25.56-1.25 1.25s.56 1.25 1.25 1.25 1.25-.56 1.25-1.25-.56-1.25-1.25-1.25zM12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2zm0 18c-4.41 0-8-3.59-8-8 0-.29.02-.58.05-.86 2.36-1.05 4.23-2.98 5.21-5.37C11.07 8.33 14.05 10 17.42 10c.78 0 1.53-.09 2.25-.26.21.71.33 1.47.33 2.26 0 4.41-3.59 8-8 8z"/></svg>';
@@ -10,6 +17,8 @@ const materialIconBase64 =
   selector: 'lab900-sharing',
   templateUrl: './sharing.component.html',
   styleUrls: ['./sharing.component.scss'],
+  standalone: true,
+  imports: [MatButtonModule, MatIconModule, MatDialogModule, MatTooltipModule],
 })
 export class Lab900SharingComponent {
   @Input()
