@@ -1,106 +1,188 @@
-import { Component, EventEmitter, HostBinding, Input, Output, ViewChild, ViewEncapsulation } from '@angular/core';
-import { TableCell } from '../../models/table-cell.model';
-import { Lab900TableCustomCellDirective } from '../../directives/table-custom-cell.directive';
-import { SortDirection } from '@angular/material/sort';
-import { Lab900Sort } from '../table/table.component';
-import { MatColumnDef } from '@angular/material/table';
+import {
+  ChangeDetectionStrategy,
+  Component,
+  ElementRef,
+  EventEmitter,
+  HostBinding,
+  inject,
+  Input,
+  OnDestroy,
+  Output,
+  TrackByFunction,
+  ViewChild,
+  ViewEncapsulation,
+} from '@angular/core';
+import { CellValueChangeEvent, TableCell } from '../../models/table-cell.model';
+import {
+  MatColumnDef,
+  MatTable,
+  MatTableModule,
+} from '@angular/material/table';
 import { readPropValue } from '../../../utils/utils';
+import {
+  BehaviorSubject,
+  combineLatest,
+  Observable,
+  ReplaySubject,
+  Subscription,
+} from 'rxjs';
+import { distinctUntilChanged, filter, map, shareReplay } from 'rxjs/operators';
+import { Lab900TableService } from '../../services/table.service';
+import { Lab900Sort } from '../../models/table-sort.model';
+import { AsyncPipe, NgClass, NgComponentOutlet } from '@angular/common';
+import { DefaultCellRendererComponent } from '../../cell-renderers/default-cell-renderer/default-cell-renderer.component';
+import { DefaultColumnHeaderRendererComponent } from '../../column-header-renderers/default-column-header-renderer/default-column-header-renderer.component';
+import { MatTooltipModule } from '@angular/material/tooltip';
+import { TranslateModule } from '@ngx-translate/core';
+import { TableCellInnerComponent } from '../table-cell-inner/table-cell-inner.component';
+import { TableCellEventsDirective } from '../../directives/table-cell-events.directive';
 
 @Component({
-  selector: 'lab900-table-cell',
+  selector: 'lab900-table-cell[cell]',
   templateUrl: './table-cell.component.html',
+  standalone: true,
+  imports: [
+    AsyncPipe,
+    DefaultCellRendererComponent,
+    MatTableModule,
+    NgClass,
+    DefaultColumnHeaderRendererComponent,
+    MatTooltipModule,
+    TranslateModule,
+    NgComponentOutlet,
+    TableCellInnerComponent,
+    TableCellEventsDirective,
+  ],
+  changeDetection: ChangeDetectionStrategy.OnPush,
   encapsulation: ViewEncapsulation.None,
 })
-export class Lab900TableCellComponent<T = any> {
+export class Lab900TableCellComponent<T = any> implements OnDestroy {
+  private cellSub: Subscription;
+  private readonly tableService = inject(Lab900TableService);
+  public readonly table = inject(MatTable);
+
   @HostBinding()
-  public className = 'lab900-table-cell';
+  public readonly className = 'lab900-table-cell';
 
   @ViewChild(MatColumnDef, { static: true })
-  public columnDef!: MatColumnDef;
+  public readonly columnDef!: MatColumnDef;
 
-  // tslint:disable-next-line:variable-name
-  private _cell: TableCell<T>;
+  @ViewChild('.lab900-td')
+  public readonly tdElement!: ElementRef<HTMLTableCellElement>;
 
-  get cell(): TableCell<T> {
-    return this._cell;
+  private readonly _cell$ = new ReplaySubject<TableCell<T>>();
+  public readonly cell$: Observable<TableCell<T>> = this._cell$
+    .asObservable()
+    .pipe(
+      filter((c) => !!c?.key),
+      shareReplay({ bufferSize: 1, refCount: true }),
+    );
+
+  private readonly tableColumnMaxWidth$ = new BehaviorSubject<string>('100%');
+
+  public readonly columnMaxWidth$: Observable<string> = combineLatest([
+    this.cell$,
+    this.tableColumnMaxWidth$,
+  ]).pipe(
+    map(
+      ([cell, tableColumnMaxWidth]) =>
+        cell?.cellMaxWidth || tableColumnMaxWidth,
+    ),
+    distinctUntilChanged(),
+    shareReplay({ bufferSize: 1, refCount: true }),
+  );
+
+  public readonly columnWidth$: Observable<string> = this.cell$.pipe(
+    map((cell) => {
+      if (cell?.width === '*') {
+        return '100%';
+      }
+      return cell?.width;
+    }),
+    distinctUntilChanged(),
+    shareReplay({ bufferSize: 1, refCount: true }),
+  );
+
+  @Input()
+  public set cell(value: TableCell<T>) {
+    this._cell$.next(value);
+  }
+
+  private readonly _data$ = new ReplaySubject<T[]>();
+
+  @Input()
+  public set data(value: T[]) {
+    this._data$.next(value);
   }
 
   @Input()
-  set cell(value: TableCell<T>) {
-    this._cell = value;
-    if (value?.key) {
-      this.columnDef.name = this._cell.key;
-      this.cellHeaderClass = this.getCellHeaderClass();
-      this.cellHeaderIcon = this.getCellHeaderIcon();
-      this.cellHeaderSvgIcon = this.getCellHeaderSvgIcon();
-      this.cellLabel = this.getCellLabel();
-    }
-  }
-
-  @Input()
-  public sort: Lab900Sort[] = [];
-
-  @Input()
-  public customCellContent?: Lab900TableCustomCellDirective;
-
-  @Input()
-  public data: T[];
+  public disableSort = false;
 
   /**
    * max column width, set by table input
    */
   @Input()
-  public maxColumnWidthFromTable?: string;
+  public set maxColumnWidthFromTable(value: string) {
+    this.tableColumnMaxWidth$.next(value);
+  }
 
   @Output()
-  public headerClick = new EventEmitter<TableCell<T>>();
+  private readonly headerClick = new EventEmitter<TableCell<T>>();
 
-  public cellClass: string;
-  public cellHeaderClass: string;
-  public cellHeaderIcon: string;
-  public cellHeaderSvgIcon: string;
-  public cellLabel: string;
+  @Output()
+  public readonly valueChanged = new EventEmitter<CellValueChangeEvent<T>>();
 
-  public get sortDirection(): SortDirection {
-    return (this.sort || []).find((s) => s.id === this.cell.key)?.direction ?? '';
-  }
+  public readonly cellHeaderClass$: Observable<string>;
+  public readonly sticky$: Observable<boolean>;
+  public readonly cellFooter$: Observable<string>;
+  public readonly sort$: Observable<Lab900Sort[] | null> =
+    this.tableService.sort$;
 
-  public get sortIcon(): string {
-    if (this.sortDirection === 'asc') {
-      return 'north';
-    } else if (this.sortDirection === 'desc') {
-      return 'south';
-    }
-  }
+  public readonly defaultCellRenderer = DefaultCellRendererComponent;
+  public readonly defaultHeaderRenderer = DefaultColumnHeaderRendererComponent;
 
-  public getCellClass(data: T): string {
-    return typeof this.cell.cellClass === 'function'
-      ? (this.cell.cellClass as (data: T, cell: TableCell) => string)(data, this.cell)
-      : this.cell.cellClass;
-  }
+  @Input({ required: true })
+  public trackByTableFn!: TrackByFunction<T>;
 
-  public getCellLabel(): string {
-    return readPropValue<TableCell<T>>(this.cell.label, this.cell);
-  }
-
-  public getCellHeaderClass(): string {
-    return readPropValue<TableCell<T>>(this.cell.cellHeaderClass, this.cell);
-  }
-
-  public getCellHeaderIcon(): string {
-    return readPropValue<TableCell<T>>(this.cell.cellHeaderIcon, this.cell);
-  }
-
-  public getCellHeaderSvgIcon(): string {
-    return readPropValue<TableCell<T>>(this.cell.cellHeaderSvgIcon, this.cell);
-  }
-
-  public getCellFooter(): string {
-    if (this.cell.footer) {
-      if (typeof this.cell.footer === 'function') {
-        return this.cell.footer(this.data, this.cell);
+  public constructor() {
+    this.cellSub = this.cell$.subscribe((cell) => {
+      if (this.columnDef.name) {
+        this.table?.removeColumnDef(this.columnDef);
       }
-      return this.cell.footer;
+      this.columnDef.name = cell.key;
+      this.table?.addColumnDef(this.columnDef);
+    });
+
+    this.sticky$ = this.cell$.pipe(map((c) => !c?.hide && !!c?.sticky));
+
+    this.cellHeaderClass$ = this.cell$.pipe(
+      map((c) => readPropValue<TableCell<T>>(c.cellHeaderClass, c)),
+    );
+
+    this.cellFooter$ = combineLatest([
+      this.cell$.pipe(filter((c) => !!c.footer)),
+      this._data$.asObservable(),
+    ]).pipe(
+      map(([cell, data]) =>
+        typeof cell.footer === 'function'
+          ? cell.footer(data, cell)
+          : cell.footer,
+      ),
+    );
+  }
+
+  public ngOnDestroy(): void {
+    this.cellSub?.unsubscribe();
+    this.table?.removeColumnDef(this.columnDef);
+  }
+
+  public handleHeaderClick(cell: TableCell<T>): void {
+    if (!this.disableSort) {
+      this.headerClick.emit(cell);
     }
+  }
+
+  public closeEditor(): void {
+    this.tableService.closeInlineEditing();
   }
 }
