@@ -1,7 +1,8 @@
 import { Directive } from '@angular/core';
 
 @Directive({
-  // tslint:disable-next-line:directive-selector
+  // eslint-disable-next-line @angular-eslint/directive-selector
   selector: '[lab900TableDisabled]',
+  standalone: true,
 })
 export class Lab900TableDisabledDirective {}
