@@ -5,71 +5,59 @@
 // common
 export * from './lib/common/models/paging.model';
 export * from './lib/utils/utils';
+export * from './lib/utils/different.utils';
 
 // dialogs
-export * from './lib/dialog/dialog.module';
 export * from './lib/dialog/components/alert-dialog/alert-dialog.component';
 export * from './lib/dialog/components/confirmation-dialog/confirmation-dialog.component';
 export * from './lib/dialog/directives/confirmation-dialog.directive';
+export * from './lib/dialog/models/alertDialog';
+export * from './lib/dialog/models/confirmationDialog';
 
 // alerts
-export * from './lib/alert/alert.module';
 export * from './lib/alert/components/alert/alert.component';
 export * from './lib/alert/models/alert';
 
 // nav list
-export * from './lib/nav-list/nav-list.module';
 export * from './lib/nav-list/models/nav-item.model';
 export * from './lib/nav-list/components/nav-list/nav-list.component';
 export * from './lib/nav-list/components/icon/icon.component';
 
 // data list
-export * from './lib/data-list/data-list.module';
 export * from './lib/data-list/models/data-list.model';
 export * from './lib/data-list/components/data-list/data-list.component';
+export * from './lib/data-list/components/data-list-item/data-list-item.component';
 export * from './lib/data-list/directives/data-list-empty.directive';
 export * from './lib/data-list/directives/data-list-item-info.directive';
 
 // sharing
-export * from './lib/sharing/sharing.module';
 export * from './lib/sharing/components/sharing/sharing.component';
 export * from './lib/sharing/models/share-dialog-component.abstract';
 
-// table
-export * from './lib/table/table.module';
-export * from './lib/table/components/table/table.component';
-export * from './lib/table/components/table-header/lab900-table-header.component';
-export * from './lib/table/components/table-filter-menu/table-filter-menu.component';
-export * from './lib/table/components/table-cell/table-cell.component';
-export * from './lib/table/components/table-cell-value/table-cell-value.component';
-export * from './lib/table/directives/table-empty.directive';
-export * from './lib/table/directives/table-custom-cell.directive';
-export * from './lib/table/directives/table-disabled.directive';
-export * from './lib/table/directives/table-header-content.directive';
-export * from './lib/table/directives/table-top-content.directive';
-export * from './lib/table/models/table-cell.model';
-export * from './lib/table/models/table-cell-tooltip.model';
-
 // page header
-export * from './lib/page-header/page-header.module';
 export * from './lib/page-header/components/page-header/page-header.component';
+export * from './lib/page-header/components/page-header-nav-item/page-header-nav-item.component';
 export * from './lib/page-header/models/page-header-nav.model';
 
 // merger
-export * from './lib/merger/merger.module';
 export * from './lib/merger/components/merger/merger.component';
+export * from './lib/merger/components/merger-item/merger-item.component';
 export * from './lib/merger/models/merge-object.model';
 export * from './lib/merger/models/merge-config.model';
 export * from './lib/merger/abstracts/custom-component.abstract';
 
 // buttons
-export * from './lib/button/button.module';
 export * from './lib/button/components/button/button.component';
 export * from './lib/button/components/action-button/lab900-action-button.component';
+export * from './lib/button/components/action-button-menu/lab900-action-button-menu.component';
+export * from './lib/button/components/action-button-toggle/lab900-action-button-toggle.component';
 export * from './lib/button/models/action-button.model';
 export * from './lib/button/models/button.model';
+export * from './lib/button/directives/preventDoubleClick.directive';
 
 // bread crumbs
-export * from './lib/bread-crumbs/bread-crumbs.module';
 export * from './lib/bread-crumbs/components/bread-crumbs/bread-crumbs.component';
 export * from './lib/bread-crumbs/models/bread-crumb.model';
+
+// table
+export * from './lib/table/table-api';
