@@ -1,16 +1,32 @@
-import { Component } from '@angular/core';
+import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { Observable } from 'rxjs';
-import { ActivatedRoute } from '@angular/router';
+import { ActivatedRoute, RouterLink } from '@angular/router';
 import { ShowcaseConfigModel } from '../../models/showcase-config.model';
 import { NavItemGroup } from '@lab900/ui';
+import { CommonModule } from '@angular/common';
+import { MatIconModule } from '@angular/material/icon';
+import { TranslateModule } from '@ngx-translate/core';
+import { MatButtonModule } from '@angular/material/button';
 
 @Component({
   selector: 'lab900-showcase-home',
   templateUrl: './showcase-home.component.html',
   styleUrls: ['./showcase-home.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: true,
+  imports: [
+    CommonModule,
+    MatIconModule,
+    TranslateModule,
+    RouterLink,
+    MatButtonModule,
+  ],
 })
 export class ShowcaseHomeComponent {
-  public data$: Observable<{ config: ShowcaseConfigModel; nav: NavItemGroup[] }> = this.activatedRoute.data as any;
+  public data$: Observable<{
+    config: ShowcaseConfigModel;
+    nav: NavItemGroup[];
+  }> = this.activatedRoute.data as any;
 
-  constructor(private activatedRoute: ActivatedRoute) {}
+  public constructor(private activatedRoute: ActivatedRoute) {}
 }
