@@ -1,3 +1,8 @@
 export class ShowcaseExample {
-  public constructor(public component: any, public title?: string, public fileFolder?: string, public extensions: string[] = ['TS']) {}
+  public constructor(
+    public component: any,
+    public title?: string,
+    public fileFolder?: string,
+    public extensions: string[] = ['TS'],
+  ) {}
 }
