@@ -1,27 +1,55 @@
 import { Component, ViewChild } from '@angular/core';
 import { mergerDataExample } from './config/merger-data-example';
-import { MergeObject } from '@lab900/ui';
-import { MergeConfig } from '@lab900/ui';
+import { Lab900MergerComponent, MergeConfig, MergeObject } from '@lab900/ui';
 import { mergerSchemaExample } from './config/merger-schema-example';
-import { Lab900MergerComponent } from '@lab900/ui';
 import { MergerDataExample } from './models/merger-data-example.model';
+import { MatButtonModule } from '@angular/material/button';
+import { TranslateModule } from '@ngx-translate/core';
 
 @Component({
   selector: 'lab900-merger-example',
+  standalone: true,
+  imports: [MatButtonModule, TranslateModule, Lab900MergerComponent],
   template: `
-    <div fxLayoutAlign="flex-end center" style="margin-bottom: 1rem">
-      <button mat-flat-button color="primary" (click)="mergerComponent.reset()">{{ 'reset' | translate }}</button>
+    <div class="merger-example-header">
+      <button
+        mat-flat-button
+        color="primary"
+        (click)="mergerComponent?.reset()"
+      >
+        {{ 'reset' | translate }}
+      </button>
     </div>
-    <lab900-merger [leftObject]="exampleData[0]" [rightObject]="exampleData[1]" [schema]="exampleSchema"></lab900-merger>
-    <button style="margin-right: 1rem" mat-raised-button color="primary" (click)="showResult()">{{ 'log_result' | translate }}</button>
+    <lab900-merger
+      [leftObject]="exampleData[0]"
+      [rightObject]="exampleData[1]"
+      [schema]="exampleSchema"
+    />
+    <button
+      style="margin-right: 1rem"
+      mat-raised-button
+      color="primary"
+      (click)="showResult()"
+    >
+      {{ 'log_result' | translate }}
+    </button>
   `,
+  styles: [
+    `
+      .merger-example-header {
+        display: flex;
+        justify-content: flex-end;
+        margin-bottom: 1rem;
+      }
+    `,
+  ],
 })
 export class MergerExampleComponent {
   public exampleData: MergeObject<MergerDataExample>[] = mergerDataExample;
   public exampleSchema: MergeConfig<MergerDataExample>[] = mergerSchemaExample;
 
   @ViewChild(Lab900MergerComponent)
-  public mergerComponent: Lab900MergerComponent<MergerDataExample>;
+  public mergerComponent?: Lab900MergerComponent<MergerDataExample>;
 
   public showResult(): void {
     console.log(this.mergerComponent.result);
