@@ -1,8 +1,15 @@
 import { Component } from '@angular/core';
+import { Lab900SharingComponent } from '@lab900/ui';
 
 @Component({
   selector: 'lab900-sharing-example',
-  template: `<lab900-sharing [users]="users" [userLabelFn]="userLabelFn" [userImageFn]="userImageFn"></lab900-sharing>`,
+  template: ` <lab900-sharing
+    [users]="users"
+    [userLabelFn]="userLabelFn"
+    [userImageFn]="userImageFn"
+  />`,
+  standalone: true,
+  imports: [Lab900SharingComponent],
 })
 export class SharingExampleComponent {
   public users = [
@@ -19,6 +26,6 @@ export class SharingExampleComponent {
     { username: 'testUser' },
   ];
 
-  public userLabelFn = (user: any) => user.username;
-  public userImageFn = (user: any) => user.profileImage || '';
+  public userLabelFn = (user: any): string => user.username;
+  public userImageFn = (user: any): string => user.profileImage || '';
 }
