@@ -1,20 +1,34 @@
 import { Component, ViewChild } from '@angular/core';
-import { Lab900Sort, Lab900TableComponent, TableCell, TableRowAction } from '@lab900/ui';
+import {
+  Lab900Sort,
+  Lab900TableComponent,
+  TableCell,
+  TableRowAction,
+} from '@lab900/ui';
 import { CdkDragDrop, moveItemInArray } from '@angular/cdk/drag-drop';
 import { SelectionModel } from '@angular/cdk/collections';
 
 @Component({
   selector: 'lab900-table-example',
+  standalone: true,
+  imports: [Lab900TableComponent],
   template: `<lab900-table
     [tableCells]="tableCells"
     [data]="mockData"
     [tableActionsBack]="tableActions"
     (tableRowOrderChange)="dropTable($event)"
-    [selectableRows]="true"
-    [selectableRowsOptions]="{ checkBoxColor: 'accent', position: 'right', sticky: true, showSelectAllCheckbox: true }"
+    [stickyHeader]="true"
+    [selectableRows]="{
+      enabled: false,
+      checkBoxColor: 'accent',
+      position: 'right',
+      sticky: true,
+      showSelectAllCheckbox: true,
+      hideSelectableRow: hideSelectableCheckboxForSarah
+    }"
     (selectionChanged)="selectionChanged($event)"
-  >
-  </lab900-table>`,
+    [trackByTableFn]="trackByTableFn"
+  />`,
 })
 export class TableDragAndDropExampleComponent {
   @ViewChild(Lab900TableComponent)
@@ -66,12 +80,15 @@ export class TableDragAndDropExampleComponent {
     {
       key: 'id',
       label: 'ID',
+      width: '50px',
     },
     {
       key: 'name',
       label: 'Name',
     },
   ];
+
+  public trackByTableFn = (index: number, item: any): any => item.id;
 
   public dropTable(event: CdkDragDrop<any[]>): void {
     /**
@@ -83,5 +100,9 @@ export class TableDragAndDropExampleComponent {
 
   public selectionChanged(selection: SelectionModel<any>): void {
     console.log('selection', selection.selected);
+  }
+
+  public hideSelectableCheckboxForSarah(row: any): boolean {
+    return row?.name === 'Sarah';
   }
 }
