@@ -1,37 +1,2846 @@
-import { Component } from '@angular/core';
-import { ActionButton, Lab900Sort, Paging, TableCell } from '@lab900/ui';
+import { Component, TrackByFunction } from '@angular/core';
+import {
+  ActionButton,
+  CellInputEditorComponent,
+  CellInputEditorOptions,
+  CellSelectEditorComponent,
+  CellSelectEditorOptions,
+  CellValueChangeEvent,
+  CellWithClearingRendererComponent,
+  Lab900Sort,
+  Lab900TableComponent,
+  Lab900TableEmptyDirective,
+  Lab900TableHeaderContentDirective,
+  Lab900TableTopContentDirective,
+  TableCell,
+} from '@lab900/ui';
+
+const mockData = [
+  {
+    uuid: '5d31c5f5-3c72-4b94-a448-008cac824707',
+    contractUuid: 'f3669b80-9237-4ab0-b107-13787bc894ab',
+    originalOrderUuid: '9229c24f-f059-435d-b659-1afda262e95a',
+    status: 'DRAFT',
+    expeditionLogStatus: 'DRAFT',
+    plannedDate: null,
+    clientReferences: null,
+    transporter: null,
+    collector: null,
+    addressee: null,
+    activity: null,
+    intermodal: null,
+    fullFreight: null,
+    nominatedQuantityBaseUnit: {
+      amount: null,
+      unit: {
+        uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+        symbol: 'MT AIR',
+      },
+    },
+    processedQuantityBaseUnit: null,
+    ediOrder: null,
+    lcsOrderFields: {
+      uuid: '1988dbba-6a9c-4d60-853f-1e47d2d64313',
+      customsStatus: 'COA_NEEDED',
+      lcsInternal: null,
+      lcsInstructionsWeighbridgeClerk: null,
+    },
+    remarks: null,
+    additionalInstructions: null,
+    plannedStockLocations: [],
+    allPlannedStockLocations: [
+      {
+        classType: 'TANK',
+        uuid: '18ed4d34-f717-4381-aaa1-198f34474401',
+        quay: {
+          uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+          id: 'ST300',
+          title: 'ST300',
+          companyId: 'ST300',
+        },
+        name: '1002',
+        tank: {
+          uuid: '8ef9000c-064c-42bf-b003-e547915eb9c2',
+          code: '1002',
+          inactive: null,
+          inactivityType: null,
+          loadingPlaces: 2,
+          loadingStation: null,
+          decisionMatrixProduct: null,
+          physicalTankQualityPercentage: null,
+          upgradePercentage: null,
+        },
+        product: {
+          uuid: '5345881c-c769-4cb0-90a8-826f81ec2347',
+          id: 'P0000038',
+          quay: {
+            uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+            id: 'ST300',
+            title: 'ST300',
+            companyId: 'ST300',
+          },
+          status: 'ACTIVE',
+          overallApproval: 'MANUALLY_RELEASED',
+          client: {
+            id: 'K0050',
+            uuid: 'cc423e20-6533-4d6e-a058-ef3aa8cf0ea2',
+            name: 'TOTSA Total Oil Trading SA',
+          },
+          mainProduct: {
+            uuid: '109a4e59-2426-4585-9908-a11e2a6c6fe7',
+            description: 'C8',
+            group: {
+              uuid: '2210753a-631f-4746-9a7b-91673f4a088c',
+              title: 'GASOLINES',
+            },
+          },
+          productProperty: {
+            uuid: 'a64f879a-53ee-4a5b-962d-dcb145b88dd4',
+            title: ' ',
+          },
+          quality: {
+            uuid: 'f4a1d27f-01da-44d3-a453-d39f80c6826e',
+            title: ' ',
+          },
+          manufacturer: {
+            uuid: '7f26e016-6002-4918-afbc-888095b2163b',
+            code: 'BP',
+            title: 'BP',
+            address: {
+              uuid: null,
+              addressLine1: 'abc',
+              addressLine2: null,
+              postalCode: null,
+              city: null,
+              country: null,
+            },
+            blockages: [],
+            types: ['CONSIGNOR', 'PRODUCT_MANUFACTURER'],
+            vatNumber: null,
+            venNumber: null,
+          },
+          baseUnit: {
+            uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+            symbol: 'MT AIR',
+            title: 'Metric Ton Air',
+          },
+          unit2: {
+            uuid: 'daea0af5-de5d-46b3-a07e-507026137a15',
+            symbol: 'm3',
+            title: 'Cubic Meter',
+          },
+          unit3: {
+            uuid: '1a76317f-5304-4187-b336-ae3fd37186be',
+            symbol: 'MT VAC',
+            title: 'Metric Ton Vaccum',
+          },
+          type: 'COMMERCIAL_PRODUCT',
+          animalByProducts: null,
+          unNumber: {
+            uuid: '595947d9-9e5a-4fb7-b685-8cb11ad70971',
+            code: '456',
+            number: '4545',
+          },
+        },
+        productGroup: null,
+        baseUnit: {
+          uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+          symbol: 'MT AIR',
+          title: 'Metric Ton Air',
+        },
+        unit2: null,
+        unit3: null,
+        limitNearlyFull: null,
+        limitNearlyEmpty: null,
+      },
+      {
+        classType: 'TANK',
+        uuid: '131f6e1b-4f27-4cae-8174-d5966f0f750c',
+        quay: {
+          uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+          id: 'ST300',
+          title: 'ST300',
+          companyId: 'ST300',
+        },
+        name: '1004',
+        tank: {
+          uuid: 'e9652e74-8de2-4d8f-9a17-456439a169fd',
+          code: '1004',
+          inactive: null,
+          inactivityType: null,
+          loadingPlaces: null,
+          loadingStation: null,
+          decisionMatrixProduct: null,
+          physicalTankQualityPercentage: null,
+          upgradePercentage: null,
+        },
+        product: {
+          uuid: '9785fd5d-6782-40e5-9af2-6b7b1e3dfe64',
+          id: 'P0000045',
+          quay: {
+            uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+            id: 'ST300',
+            title: 'ST300',
+            companyId: 'ST300',
+          },
+          status: 'ACTIVE',
+          overallApproval: 'MANUALLY_RELEASED',
+          client: {
+            id: 'K0050',
+            uuid: 'cc423e20-6533-4d6e-a058-ef3aa8cf0ea2',
+            name: 'TOTSA Total Oil Trading SA',
+          },
+          mainProduct: {
+            uuid: '09312045-f999-411a-9156-d0b9af8aef9b',
+            description: 'JET',
+            group: {
+              uuid: '7cb0785b-f02f-49fb-86a6-45dd836a6faa',
+              title: 'JET',
+            },
+          },
+          productProperty: {
+            uuid: 'a64f879a-53ee-4a5b-962d-dcb145b88dd4',
+            title: ' ',
+          },
+          quality: {
+            uuid: 'f4a1d27f-01da-44d3-a453-d39f80c6826e',
+            title: ' ',
+          },
+          manufacturer: {
+            uuid: '580e9c9f-e2a6-4c6f-834f-55c393ddb285',
+            code: 'ASTRO BIOPLANT',
+            title: 'Astro Bioplant',
+            address: null,
+            blockages: [],
+            types: ['PRODUCT_MANUFACTURER'],
+            vatNumber: null,
+            venNumber: null,
+          },
+          baseUnit: {
+            uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+            symbol: 'MT AIR',
+            title: 'Metric Ton Air',
+          },
+          unit2: {
+            uuid: 'daea0af5-de5d-46b3-a07e-507026137a15',
+            symbol: 'm3',
+            title: 'Cubic Meter',
+          },
+          unit3: {
+            uuid: '1a76317f-5304-4187-b336-ae3fd37186be',
+            symbol: 'MT VAC',
+            title: 'Metric Ton Vaccum',
+          },
+          type: 'COMMERCIAL_PRODUCT',
+          animalByProducts: null,
+          unNumber: null,
+        },
+        productGroup: null,
+        baseUnit: {
+          uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+          symbol: 'MT AIR',
+          title: 'Metric Ton Air',
+        },
+        unit2: null,
+        unit3: null,
+        limitNearlyFull: null,
+        limitNearlyEmpty: null,
+      },
+    ],
+    sofStockLocationNames: [],
+    wmsStatus: null,
+    invoicingCode: null,
+    createdOn: '2024-03-25T08:51:37.5117058Z',
+  },
+  {
+    uuid: '9229c24f-f059-435d-b659-1afda262e95a',
+    contractUuid: 'f3669b80-9237-4ab0-b107-13787bc894ab',
+    originalOrderUuid: null,
+    status: 'DRAFT',
+    expeditionLogStatus: 'DRAFT',
+    plannedDate: null,
+    clientReferences: {
+      reference1: 'abc102',
+      reference2: null,
+      reference3: null,
+    },
+    transporter: null,
+    collector: null,
+    addressee: null,
+    activity: null,
+    intermodal: null,
+    fullFreight: null,
+    nominatedQuantityBaseUnit: {
+      amount: null,
+      unit: {
+        uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+        symbol: 'MT AIR',
+      },
+    },
+    processedQuantityBaseUnit: null,
+    ediOrder: null,
+    lcsOrderFields: {
+      uuid: '2ac34abb-696f-4654-a53b-7c9efb47083b',
+      customsStatus: null,
+      lcsInternal: null,
+      lcsInstructionsWeighbridgeClerk: null,
+    },
+    remarks: null,
+    additionalInstructions: null,
+    plannedStockLocations: [],
+    allPlannedStockLocations: [
+      {
+        classType: 'TANK',
+        uuid: '18ed4d34-f717-4381-aaa1-198f34474401',
+        quay: {
+          uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+          id: 'ST300',
+          title: 'ST300',
+          companyId: 'ST300',
+        },
+        name: '1002',
+        tank: {
+          uuid: '8ef9000c-064c-42bf-b003-e547915eb9c2',
+          code: '1002',
+          inactive: null,
+          inactivityType: null,
+          loadingPlaces: 2,
+          loadingStation: null,
+          decisionMatrixProduct: null,
+          physicalTankQualityPercentage: null,
+          upgradePercentage: null,
+        },
+        product: {
+          uuid: '5345881c-c769-4cb0-90a8-826f81ec2347',
+          id: 'P0000038',
+          quay: {
+            uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+            id: 'ST300',
+            title: 'ST300',
+            companyId: 'ST300',
+          },
+          status: 'ACTIVE',
+          overallApproval: 'MANUALLY_RELEASED',
+          client: {
+            id: 'K0050',
+            uuid: 'cc423e20-6533-4d6e-a058-ef3aa8cf0ea2',
+            name: 'TOTSA Total Oil Trading SA',
+          },
+          mainProduct: {
+            uuid: '109a4e59-2426-4585-9908-a11e2a6c6fe7',
+            description: 'C8',
+            group: {
+              uuid: '2210753a-631f-4746-9a7b-91673f4a088c',
+              title: 'GASOLINES',
+            },
+          },
+          productProperty: {
+            uuid: 'a64f879a-53ee-4a5b-962d-dcb145b88dd4',
+            title: ' ',
+          },
+          quality: {
+            uuid: 'f4a1d27f-01da-44d3-a453-d39f80c6826e',
+            title: ' ',
+          },
+          manufacturer: {
+            uuid: '7f26e016-6002-4918-afbc-888095b2163b',
+            code: 'BP',
+            title: 'BP',
+            address: {
+              uuid: null,
+              addressLine1: 'abc',
+              addressLine2: null,
+              postalCode: null,
+              city: null,
+              country: null,
+            },
+            blockages: [],
+            types: ['CONSIGNOR', 'PRODUCT_MANUFACTURER'],
+            vatNumber: null,
+            venNumber: null,
+          },
+          baseUnit: {
+            uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+            symbol: 'MT AIR',
+            title: 'Metric Ton Air',
+          },
+          unit2: {
+            uuid: 'daea0af5-de5d-46b3-a07e-507026137a15',
+            symbol: 'm3',
+            title: 'Cubic Meter',
+          },
+          unit3: {
+            uuid: '1a76317f-5304-4187-b336-ae3fd37186be',
+            symbol: 'MT VAC',
+            title: 'Metric Ton Vaccum',
+          },
+          type: 'COMMERCIAL_PRODUCT',
+          animalByProducts: null,
+          unNumber: {
+            uuid: '595947d9-9e5a-4fb7-b685-8cb11ad70971',
+            code: '456',
+            number: '4545',
+          },
+        },
+        productGroup: null,
+        baseUnit: {
+          uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+          symbol: 'MT AIR',
+          title: 'Metric Ton Air',
+        },
+        unit2: null,
+        unit3: null,
+        limitNearlyFull: null,
+        limitNearlyEmpty: null,
+      },
+      {
+        classType: 'TANK',
+        uuid: '131f6e1b-4f27-4cae-8174-d5966f0f750c',
+        quay: {
+          uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+          id: 'ST300',
+          title: 'ST300',
+          companyId: 'ST300',
+        },
+        name: '1004',
+        tank: {
+          uuid: 'e9652e74-8de2-4d8f-9a17-456439a169fd',
+          code: '1004',
+          inactive: null,
+          inactivityType: null,
+          loadingPlaces: null,
+          loadingStation: null,
+          decisionMatrixProduct: null,
+          physicalTankQualityPercentage: null,
+          upgradePercentage: null,
+        },
+        product: {
+          uuid: '9785fd5d-6782-40e5-9af2-6b7b1e3dfe64',
+          id: 'P0000045',
+          quay: {
+            uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+            id: 'ST300',
+            title: 'ST300',
+            companyId: 'ST300',
+          },
+          status: 'ACTIVE',
+          overallApproval: 'MANUALLY_RELEASED',
+          client: {
+            id: 'K0050',
+            uuid: 'cc423e20-6533-4d6e-a058-ef3aa8cf0ea2',
+            name: 'TOTSA Total Oil Trading SA',
+          },
+          mainProduct: {
+            uuid: '09312045-f999-411a-9156-d0b9af8aef9b',
+            description: 'JET',
+            group: {
+              uuid: '7cb0785b-f02f-49fb-86a6-45dd836a6faa',
+              title: 'JET',
+            },
+          },
+          productProperty: {
+            uuid: 'a64f879a-53ee-4a5b-962d-dcb145b88dd4',
+            title: ' ',
+          },
+          quality: {
+            uuid: 'f4a1d27f-01da-44d3-a453-d39f80c6826e',
+            title: ' ',
+          },
+          manufacturer: {
+            uuid: '580e9c9f-e2a6-4c6f-834f-55c393ddb285',
+            code: 'ASTRO BIOPLANT',
+            title: 'Astro Bioplant',
+            address: null,
+            blockages: [],
+            types: ['PRODUCT_MANUFACTURER'],
+            vatNumber: null,
+            venNumber: null,
+          },
+          baseUnit: {
+            uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+            symbol: 'MT AIR',
+            title: 'Metric Ton Air',
+          },
+          unit2: {
+            uuid: 'daea0af5-de5d-46b3-a07e-507026137a15',
+            symbol: 'm3',
+            title: 'Cubic Meter',
+          },
+          unit3: {
+            uuid: '1a76317f-5304-4187-b336-ae3fd37186be',
+            symbol: 'MT VAC',
+            title: 'Metric Ton Vaccum',
+          },
+          type: 'COMMERCIAL_PRODUCT',
+          animalByProducts: null,
+          unNumber: null,
+        },
+        productGroup: null,
+        baseUnit: {
+          uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+          symbol: 'MT AIR',
+          title: 'Metric Ton Air',
+        },
+        unit2: null,
+        unit3: null,
+        limitNearlyFull: null,
+        limitNearlyEmpty: null,
+      },
+    ],
+    sofStockLocationNames: [],
+    wmsStatus: null,
+    invoicingCode: {
+      uuid: 'e0b608c4-a46a-4c9b-8b37-21a16817d6b4',
+      code: 'JET',
+    },
+    createdOn: '2024-03-25T08:47:36.2872395Z',
+  },
+  {
+    uuid: '4e0dfdbe-8d2b-46b6-9b59-8589a6869c9e',
+    contractUuid: 'f3669b80-9237-4ab0-b107-13787bc894ab',
+    originalOrderUuid: null,
+    status: 'DRAFT',
+    expeditionLogStatus: 'DRAFT',
+    plannedDate: null,
+    clientReferences: {
+      reference1: 'abc101',
+      reference2: null,
+      reference3: null,
+    },
+    transporter: null,
+    collector: null,
+    addressee: null,
+    activity: null,
+    intermodal: null,
+    fullFreight: null,
+    nominatedQuantityBaseUnit: {
+      amount: null,
+      unit: {
+        uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+        symbol: 'MT AIR',
+      },
+    },
+    processedQuantityBaseUnit: null,
+    ediOrder: null,
+    lcsOrderFields: {
+      uuid: '7584f869-eac5-45b8-9d17-a1fa08870237',
+      customsStatus: null,
+      lcsInternal: null,
+      lcsInstructionsWeighbridgeClerk: null,
+    },
+    remarks: null,
+    additionalInstructions: null,
+    plannedStockLocations: [],
+    allPlannedStockLocations: [
+      {
+        classType: 'TANK',
+        uuid: '18ed4d34-f717-4381-aaa1-198f34474401',
+        quay: {
+          uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+          id: 'ST300',
+          title: 'ST300',
+          companyId: 'ST300',
+        },
+        name: '1002',
+        tank: {
+          uuid: '8ef9000c-064c-42bf-b003-e547915eb9c2',
+          code: '1002',
+          inactive: null,
+          inactivityType: null,
+          loadingPlaces: 2,
+          loadingStation: null,
+          decisionMatrixProduct: null,
+          physicalTankQualityPercentage: null,
+          upgradePercentage: null,
+        },
+        product: {
+          uuid: '5345881c-c769-4cb0-90a8-826f81ec2347',
+          id: 'P0000038',
+          quay: {
+            uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+            id: 'ST300',
+            title: 'ST300',
+            companyId: 'ST300',
+          },
+          status: 'ACTIVE',
+          overallApproval: 'MANUALLY_RELEASED',
+          client: {
+            id: 'K0050',
+            uuid: 'cc423e20-6533-4d6e-a058-ef3aa8cf0ea2',
+            name: 'TOTSA Total Oil Trading SA',
+          },
+          mainProduct: {
+            uuid: '109a4e59-2426-4585-9908-a11e2a6c6fe7',
+            description: 'C8',
+            group: {
+              uuid: '2210753a-631f-4746-9a7b-91673f4a088c',
+              title: 'GASOLINES',
+            },
+          },
+          productProperty: {
+            uuid: 'a64f879a-53ee-4a5b-962d-dcb145b88dd4',
+            title: ' ',
+          },
+          quality: {
+            uuid: 'f4a1d27f-01da-44d3-a453-d39f80c6826e',
+            title: ' ',
+          },
+          manufacturer: {
+            uuid: '7f26e016-6002-4918-afbc-888095b2163b',
+            code: 'BP',
+            title: 'BP',
+            address: {
+              uuid: null,
+              addressLine1: 'abc',
+              addressLine2: null,
+              postalCode: null,
+              city: null,
+              country: null,
+            },
+            blockages: [],
+            types: ['CONSIGNOR', 'PRODUCT_MANUFACTURER'],
+            vatNumber: null,
+            venNumber: null,
+          },
+          baseUnit: {
+            uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+            symbol: 'MT AIR',
+            title: 'Metric Ton Air',
+          },
+          unit2: {
+            uuid: 'daea0af5-de5d-46b3-a07e-507026137a15',
+            symbol: 'm3',
+            title: 'Cubic Meter',
+          },
+          unit3: {
+            uuid: '1a76317f-5304-4187-b336-ae3fd37186be',
+            symbol: 'MT VAC',
+            title: 'Metric Ton Vaccum',
+          },
+          type: 'COMMERCIAL_PRODUCT',
+          animalByProducts: null,
+          unNumber: {
+            uuid: '595947d9-9e5a-4fb7-b685-8cb11ad70971',
+            code: '456',
+            number: '4545',
+          },
+        },
+        productGroup: null,
+        baseUnit: {
+          uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+          symbol: 'MT AIR',
+          title: 'Metric Ton Air',
+        },
+        unit2: null,
+        unit3: null,
+        limitNearlyFull: null,
+        limitNearlyEmpty: null,
+      },
+      {
+        classType: 'TANK',
+        uuid: '131f6e1b-4f27-4cae-8174-d5966f0f750c',
+        quay: {
+          uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+          id: 'ST300',
+          title: 'ST300',
+          companyId: 'ST300',
+        },
+        name: '1004',
+        tank: {
+          uuid: 'e9652e74-8de2-4d8f-9a17-456439a169fd',
+          code: '1004',
+          inactive: null,
+          inactivityType: null,
+          loadingPlaces: null,
+          loadingStation: null,
+          decisionMatrixProduct: null,
+          physicalTankQualityPercentage: null,
+          upgradePercentage: null,
+        },
+        product: {
+          uuid: '9785fd5d-6782-40e5-9af2-6b7b1e3dfe64',
+          id: 'P0000045',
+          quay: {
+            uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+            id: 'ST300',
+            title: 'ST300',
+            companyId: 'ST300',
+          },
+          status: 'ACTIVE',
+          overallApproval: 'MANUALLY_RELEASED',
+          client: {
+            id: 'K0050',
+            uuid: 'cc423e20-6533-4d6e-a058-ef3aa8cf0ea2',
+            name: 'TOTSA Total Oil Trading SA',
+          },
+          mainProduct: {
+            uuid: '09312045-f999-411a-9156-d0b9af8aef9b',
+            description: 'JET',
+            group: {
+              uuid: '7cb0785b-f02f-49fb-86a6-45dd836a6faa',
+              title: 'JET',
+            },
+          },
+          productProperty: {
+            uuid: 'a64f879a-53ee-4a5b-962d-dcb145b88dd4',
+            title: ' ',
+          },
+          quality: {
+            uuid: 'f4a1d27f-01da-44d3-a453-d39f80c6826e',
+            title: ' ',
+          },
+          manufacturer: {
+            uuid: '580e9c9f-e2a6-4c6f-834f-55c393ddb285',
+            code: 'ASTRO BIOPLANT',
+            title: 'Astro Bioplant',
+            address: null,
+            blockages: [],
+            types: ['PRODUCT_MANUFACTURER'],
+            vatNumber: null,
+            venNumber: null,
+          },
+          baseUnit: {
+            uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+            symbol: 'MT AIR',
+            title: 'Metric Ton Air',
+          },
+          unit2: {
+            uuid: 'daea0af5-de5d-46b3-a07e-507026137a15',
+            symbol: 'm3',
+            title: 'Cubic Meter',
+          },
+          unit3: {
+            uuid: '1a76317f-5304-4187-b336-ae3fd37186be',
+            symbol: 'MT VAC',
+            title: 'Metric Ton Vaccum',
+          },
+          type: 'COMMERCIAL_PRODUCT',
+          animalByProducts: null,
+          unNumber: null,
+        },
+        productGroup: null,
+        baseUnit: {
+          uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+          symbol: 'MT AIR',
+          title: 'Metric Ton Air',
+        },
+        unit2: null,
+        unit3: null,
+        limitNearlyFull: null,
+        limitNearlyEmpty: null,
+      },
+    ],
+    sofStockLocationNames: [],
+    wmsStatus: null,
+    invoicingCode: null,
+    createdOn: '2024-03-25T08:47:36.2287243Z',
+  },
+  {
+    uuid: 'b36ce1a0-f049-4eb6-82cf-13a597fb2fd0',
+    contractUuid: 'f3669b80-9237-4ab0-b107-13787bc894ab',
+    originalOrderUuid: null,
+    status: 'DRAFT',
+    expeditionLogStatus: 'DRAFT',
+    plannedDate: null,
+    clientReferences: {
+      reference1: 'abc100',
+      reference2: null,
+      reference3: null,
+    },
+    transporter: null,
+    collector: null,
+    addressee: null,
+    activity: null,
+    intermodal: null,
+    fullFreight: null,
+    nominatedQuantityBaseUnit: {
+      amount: null,
+      unit: {
+        uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+        symbol: 'MT AIR',
+      },
+    },
+    processedQuantityBaseUnit: null,
+    ediOrder: null,
+    lcsOrderFields: {
+      uuid: 'b54c7e8d-52c2-4fef-991c-ea83af5cec8b',
+      customsStatus: null,
+      lcsInternal: null,
+      lcsInstructionsWeighbridgeClerk: null,
+    },
+    remarks: null,
+    additionalInstructions: null,
+    plannedStockLocations: [],
+    allPlannedStockLocations: [
+      {
+        classType: 'TANK',
+        uuid: '18ed4d34-f717-4381-aaa1-198f34474401',
+        quay: {
+          uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+          id: 'ST300',
+          title: 'ST300',
+          companyId: 'ST300',
+        },
+        name: '1002',
+        tank: {
+          uuid: '8ef9000c-064c-42bf-b003-e547915eb9c2',
+          code: '1002',
+          inactive: null,
+          inactivityType: null,
+          loadingPlaces: 2,
+          loadingStation: null,
+          decisionMatrixProduct: null,
+          physicalTankQualityPercentage: null,
+          upgradePercentage: null,
+        },
+        product: {
+          uuid: '5345881c-c769-4cb0-90a8-826f81ec2347',
+          id: 'P0000038',
+          quay: {
+            uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+            id: 'ST300',
+            title: 'ST300',
+            companyId: 'ST300',
+          },
+          status: 'ACTIVE',
+          overallApproval: 'MANUALLY_RELEASED',
+          client: {
+            id: 'K0050',
+            uuid: 'cc423e20-6533-4d6e-a058-ef3aa8cf0ea2',
+            name: 'TOTSA Total Oil Trading SA',
+          },
+          mainProduct: {
+            uuid: '109a4e59-2426-4585-9908-a11e2a6c6fe7',
+            description: 'C8',
+            group: {
+              uuid: '2210753a-631f-4746-9a7b-91673f4a088c',
+              title: 'GASOLINES',
+            },
+          },
+          productProperty: {
+            uuid: 'a64f879a-53ee-4a5b-962d-dcb145b88dd4',
+            title: ' ',
+          },
+          quality: {
+            uuid: 'f4a1d27f-01da-44d3-a453-d39f80c6826e',
+            title: ' ',
+          },
+          manufacturer: {
+            uuid: '7f26e016-6002-4918-afbc-888095b2163b',
+            code: 'BP',
+            title: 'BP',
+            address: {
+              uuid: null,
+              addressLine1: 'abc',
+              addressLine2: null,
+              postalCode: null,
+              city: null,
+              country: null,
+            },
+            blockages: [],
+            types: ['CONSIGNOR', 'PRODUCT_MANUFACTURER'],
+            vatNumber: null,
+            venNumber: null,
+          },
+          baseUnit: {
+            uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+            symbol: 'MT AIR',
+            title: 'Metric Ton Air',
+          },
+          unit2: {
+            uuid: 'daea0af5-de5d-46b3-a07e-507026137a15',
+            symbol: 'm3',
+            title: 'Cubic Meter',
+          },
+          unit3: {
+            uuid: '1a76317f-5304-4187-b336-ae3fd37186be',
+            symbol: 'MT VAC',
+            title: 'Metric Ton Vaccum',
+          },
+          type: 'COMMERCIAL_PRODUCT',
+          animalByProducts: null,
+          unNumber: {
+            uuid: '595947d9-9e5a-4fb7-b685-8cb11ad70971',
+            code: '456',
+            number: '4545',
+          },
+        },
+        productGroup: null,
+        baseUnit: {
+          uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+          symbol: 'MT AIR',
+          title: 'Metric Ton Air',
+        },
+        unit2: null,
+        unit3: null,
+        limitNearlyFull: null,
+        limitNearlyEmpty: null,
+      },
+      {
+        classType: 'TANK',
+        uuid: '131f6e1b-4f27-4cae-8174-d5966f0f750c',
+        quay: {
+          uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+          id: 'ST300',
+          title: 'ST300',
+          companyId: 'ST300',
+        },
+        name: '1004',
+        tank: {
+          uuid: 'e9652e74-8de2-4d8f-9a17-456439a169fd',
+          code: '1004',
+          inactive: null,
+          inactivityType: null,
+          loadingPlaces: null,
+          loadingStation: null,
+          decisionMatrixProduct: null,
+          physicalTankQualityPercentage: null,
+          upgradePercentage: null,
+        },
+        product: {
+          uuid: '9785fd5d-6782-40e5-9af2-6b7b1e3dfe64',
+          id: 'P0000045',
+          quay: {
+            uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+            id: 'ST300',
+            title: 'ST300',
+            companyId: 'ST300',
+          },
+          status: 'ACTIVE',
+          overallApproval: 'MANUALLY_RELEASED',
+          client: {
+            id: 'K0050',
+            uuid: 'cc423e20-6533-4d6e-a058-ef3aa8cf0ea2',
+            name: 'TOTSA Total Oil Trading SA',
+          },
+          mainProduct: {
+            uuid: '09312045-f999-411a-9156-d0b9af8aef9b',
+            description: 'JET',
+            group: {
+              uuid: '7cb0785b-f02f-49fb-86a6-45dd836a6faa',
+              title: 'JET',
+            },
+          },
+          productProperty: {
+            uuid: 'a64f879a-53ee-4a5b-962d-dcb145b88dd4',
+            title: ' ',
+          },
+          quality: {
+            uuid: 'f4a1d27f-01da-44d3-a453-d39f80c6826e',
+            title: ' ',
+          },
+          manufacturer: {
+            uuid: '580e9c9f-e2a6-4c6f-834f-55c393ddb285',
+            code: 'ASTRO BIOPLANT',
+            title: 'Astro Bioplant',
+            address: null,
+            blockages: [],
+            types: ['PRODUCT_MANUFACTURER'],
+            vatNumber: null,
+            venNumber: null,
+          },
+          baseUnit: {
+            uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+            symbol: 'MT AIR',
+            title: 'Metric Ton Air',
+          },
+          unit2: {
+            uuid: 'daea0af5-de5d-46b3-a07e-507026137a15',
+            symbol: 'm3',
+            title: 'Cubic Meter',
+          },
+          unit3: {
+            uuid: '1a76317f-5304-4187-b336-ae3fd37186be',
+            symbol: 'MT VAC',
+            title: 'Metric Ton Vaccum',
+          },
+          type: 'COMMERCIAL_PRODUCT',
+          animalByProducts: null,
+          unNumber: null,
+        },
+        productGroup: null,
+        baseUnit: {
+          uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+          symbol: 'MT AIR',
+          title: 'Metric Ton Air',
+        },
+        unit2: null,
+        unit3: null,
+        limitNearlyFull: null,
+        limitNearlyEmpty: null,
+      },
+    ],
+    sofStockLocationNames: [],
+    wmsStatus: null,
+    invoicingCode: null,
+    createdOn: '2024-03-25T08:47:36.1129494Z',
+  },
+  {
+    uuid: 'a428c4c6-a974-4ec0-ae1c-4c9e096b07f8',
+    contractUuid: 'f3669b80-9237-4ab0-b107-13787bc894ab',
+    originalOrderUuid: null,
+    status: 'DRAFT',
+    expeditionLogStatus: 'DRAFT',
+    plannedDate: null,
+    clientReferences: {
+      reference1: 'test',
+      reference2: '',
+      reference3: 'ezrezr',
+    },
+    transporter: null,
+    collector: null,
+    addressee: null,
+    activity: null,
+    intermodal: null,
+    fullFreight: null,
+    nominatedQuantityBaseUnit: {
+      amount: 0.01,
+      unit: {
+        uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+        symbol: 'MT AIR',
+      },
+    },
+    processedQuantityBaseUnit: null,
+    ediOrder: null,
+    lcsOrderFields: {
+      uuid: '11e0ab28-6aa0-4a90-ad31-dfffaa235784',
+      customsStatus: null,
+      lcsInternal: null,
+      lcsInstructionsWeighbridgeClerk: null,
+    },
+    remarks: null,
+    additionalInstructions: null,
+    plannedStockLocations: [],
+    allPlannedStockLocations: [
+      {
+        classType: 'TANK',
+        uuid: '18ed4d34-f717-4381-aaa1-198f34474401',
+        quay: {
+          uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+          id: 'ST300',
+          title: 'ST300',
+          companyId: 'ST300',
+        },
+        name: '1002',
+        tank: {
+          uuid: '8ef9000c-064c-42bf-b003-e547915eb9c2',
+          code: '1002',
+          inactive: null,
+          inactivityType: null,
+          loadingPlaces: 2,
+          loadingStation: null,
+          decisionMatrixProduct: null,
+          physicalTankQualityPercentage: null,
+          upgradePercentage: null,
+        },
+        product: {
+          uuid: '5345881c-c769-4cb0-90a8-826f81ec2347',
+          id: 'P0000038',
+          quay: {
+            uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+            id: 'ST300',
+            title: 'ST300',
+            companyId: 'ST300',
+          },
+          status: 'ACTIVE',
+          overallApproval: 'MANUALLY_RELEASED',
+          client: {
+            id: 'K0050',
+            uuid: 'cc423e20-6533-4d6e-a058-ef3aa8cf0ea2',
+            name: 'TOTSA Total Oil Trading SA',
+          },
+          mainProduct: {
+            uuid: '109a4e59-2426-4585-9908-a11e2a6c6fe7',
+            description: 'C8',
+            group: {
+              uuid: '2210753a-631f-4746-9a7b-91673f4a088c',
+              title: 'GASOLINES',
+            },
+          },
+          productProperty: {
+            uuid: 'a64f879a-53ee-4a5b-962d-dcb145b88dd4',
+            title: ' ',
+          },
+          quality: {
+            uuid: 'f4a1d27f-01da-44d3-a453-d39f80c6826e',
+            title: ' ',
+          },
+          manufacturer: {
+            uuid: '7f26e016-6002-4918-afbc-888095b2163b',
+            code: 'BP',
+            title: 'BP',
+            address: {
+              uuid: null,
+              addressLine1: 'abc',
+              addressLine2: null,
+              postalCode: null,
+              city: null,
+              country: null,
+            },
+            blockages: [],
+            types: ['CONSIGNOR', 'PRODUCT_MANUFACTURER'],
+            vatNumber: null,
+            venNumber: null,
+          },
+          baseUnit: {
+            uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+            symbol: 'MT AIR',
+            title: 'Metric Ton Air',
+          },
+          unit2: {
+            uuid: 'daea0af5-de5d-46b3-a07e-507026137a15',
+            symbol: 'm3',
+            title: 'Cubic Meter',
+          },
+          unit3: {
+            uuid: '1a76317f-5304-4187-b336-ae3fd37186be',
+            symbol: 'MT VAC',
+            title: 'Metric Ton Vaccum',
+          },
+          type: 'COMMERCIAL_PRODUCT',
+          animalByProducts: null,
+          unNumber: {
+            uuid: '595947d9-9e5a-4fb7-b685-8cb11ad70971',
+            code: '456',
+            number: '4545',
+          },
+        },
+        productGroup: null,
+        baseUnit: {
+          uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+          symbol: 'MT AIR',
+          title: 'Metric Ton Air',
+        },
+        unit2: null,
+        unit3: null,
+        limitNearlyFull: null,
+        limitNearlyEmpty: null,
+      },
+      {
+        classType: 'TANK',
+        uuid: '131f6e1b-4f27-4cae-8174-d5966f0f750c',
+        quay: {
+          uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+          id: 'ST300',
+          title: 'ST300',
+          companyId: 'ST300',
+        },
+        name: '1004',
+        tank: {
+          uuid: 'e9652e74-8de2-4d8f-9a17-456439a169fd',
+          code: '1004',
+          inactive: null,
+          inactivityType: null,
+          loadingPlaces: null,
+          loadingStation: null,
+          decisionMatrixProduct: null,
+          physicalTankQualityPercentage: null,
+          upgradePercentage: null,
+        },
+        product: {
+          uuid: '9785fd5d-6782-40e5-9af2-6b7b1e3dfe64',
+          id: 'P0000045',
+          quay: {
+            uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+            id: 'ST300',
+            title: 'ST300',
+            companyId: 'ST300',
+          },
+          status: 'ACTIVE',
+          overallApproval: 'MANUALLY_RELEASED',
+          client: {
+            id: 'K0050',
+            uuid: 'cc423e20-6533-4d6e-a058-ef3aa8cf0ea2',
+            name: 'TOTSA Total Oil Trading SA',
+          },
+          mainProduct: {
+            uuid: '09312045-f999-411a-9156-d0b9af8aef9b',
+            description: 'JET',
+            group: {
+              uuid: '7cb0785b-f02f-49fb-86a6-45dd836a6faa',
+              title: 'JET',
+            },
+          },
+          productProperty: {
+            uuid: 'a64f879a-53ee-4a5b-962d-dcb145b88dd4',
+            title: ' ',
+          },
+          quality: {
+            uuid: 'f4a1d27f-01da-44d3-a453-d39f80c6826e',
+            title: ' ',
+          },
+          manufacturer: {
+            uuid: '580e9c9f-e2a6-4c6f-834f-55c393ddb285',
+            code: 'ASTRO BIOPLANT',
+            title: 'Astro Bioplant',
+            address: null,
+            blockages: [],
+            types: ['PRODUCT_MANUFACTURER'],
+            vatNumber: null,
+            venNumber: null,
+          },
+          baseUnit: {
+            uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+            symbol: 'MT AIR',
+            title: 'Metric Ton Air',
+          },
+          unit2: {
+            uuid: 'daea0af5-de5d-46b3-a07e-507026137a15',
+            symbol: 'm3',
+            title: 'Cubic Meter',
+          },
+          unit3: {
+            uuid: '1a76317f-5304-4187-b336-ae3fd37186be',
+            symbol: 'MT VAC',
+            title: 'Metric Ton Vaccum',
+          },
+          type: 'COMMERCIAL_PRODUCT',
+          animalByProducts: null,
+          unNumber: null,
+        },
+        productGroup: null,
+        baseUnit: {
+          uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+          symbol: 'MT AIR',
+          title: 'Metric Ton Air',
+        },
+        unit2: null,
+        unit3: null,
+        limitNearlyFull: null,
+        limitNearlyEmpty: null,
+      },
+    ],
+    sofStockLocationNames: [],
+    wmsStatus: null,
+    invoicingCode: {
+      uuid: '9ca528ae-b1a2-4075-9790-cfec4b4d264f',
+      code: 'RENEWABLES',
+    },
+    createdOn: '2024-01-29T08:32:26.6940423Z',
+  },
+  {
+    uuid: '6ec7b814-b31c-417e-ab7a-c913a123370e',
+    contractUuid: 'f3669b80-9237-4ab0-b107-13787bc894ab',
+    originalOrderUuid: null,
+    status: 'DRAFT',
+    expeditionLogStatus: 'DRAFT',
+    plannedDate: '2023-12-27',
+    clientReferences: {
+      reference1: '',
+      reference2: '',
+      reference3: null,
+    },
+    transporter: null,
+    collector: null,
+    addressee: null,
+    activity: null,
+    intermodal: null,
+    fullFreight: null,
+    nominatedQuantityBaseUnit: {
+      amount: null,
+      unit: {
+        uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+        symbol: 'MT AIR',
+      },
+    },
+    processedQuantityBaseUnit: null,
+    ediOrder: {
+      referenceOd: '',
+      referenceId: '',
+      referenceOs: '',
+      referencePoId: '',
+      referencePoLine: '',
+    },
+    lcsOrderFields: {
+      uuid: '1f786f94-c6f7-4d29-b147-42ddd89007dd',
+      customsStatus: null,
+      lcsInternal: null,
+      lcsInstructionsWeighbridgeClerk: null,
+    },
+    remarks: {
+      changes: [],
+      customerService: '',
+      internal: '',
+    },
+    additionalInstructions: null,
+    plannedStockLocations: [],
+    allPlannedStockLocations: [
+      {
+        classType: 'TANK',
+        uuid: '18ed4d34-f717-4381-aaa1-198f34474401',
+        quay: {
+          uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+          id: 'ST300',
+          title: 'ST300',
+          companyId: 'ST300',
+        },
+        name: '1002',
+        tank: {
+          uuid: '8ef9000c-064c-42bf-b003-e547915eb9c2',
+          code: '1002',
+          inactive: null,
+          inactivityType: null,
+          loadingPlaces: 2,
+          loadingStation: null,
+          decisionMatrixProduct: null,
+          physicalTankQualityPercentage: null,
+          upgradePercentage: null,
+        },
+        product: {
+          uuid: '5345881c-c769-4cb0-90a8-826f81ec2347',
+          id: 'P0000038',
+          quay: {
+            uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+            id: 'ST300',
+            title: 'ST300',
+            companyId: 'ST300',
+          },
+          status: 'ACTIVE',
+          overallApproval: 'MANUALLY_RELEASED',
+          client: {
+            id: 'K0050',
+            uuid: 'cc423e20-6533-4d6e-a058-ef3aa8cf0ea2',
+            name: 'TOTSA Total Oil Trading SA',
+          },
+          mainProduct: {
+            uuid: '109a4e59-2426-4585-9908-a11e2a6c6fe7',
+            description: 'C8',
+            group: {
+              uuid: '2210753a-631f-4746-9a7b-91673f4a088c',
+              title: 'GASOLINES',
+            },
+          },
+          productProperty: {
+            uuid: 'a64f879a-53ee-4a5b-962d-dcb145b88dd4',
+            title: ' ',
+          },
+          quality: {
+            uuid: 'f4a1d27f-01da-44d3-a453-d39f80c6826e',
+            title: ' ',
+          },
+          manufacturer: {
+            uuid: '7f26e016-6002-4918-afbc-888095b2163b',
+            code: 'BP',
+            title: 'BP',
+            address: {
+              uuid: null,
+              addressLine1: 'abc',
+              addressLine2: null,
+              postalCode: null,
+              city: null,
+              country: null,
+            },
+            blockages: [],
+            types: ['CONSIGNOR', 'PRODUCT_MANUFACTURER'],
+            vatNumber: null,
+            venNumber: null,
+          },
+          baseUnit: {
+            uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+            symbol: 'MT AIR',
+            title: 'Metric Ton Air',
+          },
+          unit2: {
+            uuid: 'daea0af5-de5d-46b3-a07e-507026137a15',
+            symbol: 'm3',
+            title: 'Cubic Meter',
+          },
+          unit3: {
+            uuid: '1a76317f-5304-4187-b336-ae3fd37186be',
+            symbol: 'MT VAC',
+            title: 'Metric Ton Vaccum',
+          },
+          type: 'COMMERCIAL_PRODUCT',
+          animalByProducts: null,
+          unNumber: {
+            uuid: '595947d9-9e5a-4fb7-b685-8cb11ad70971',
+            code: '456',
+            number: '4545',
+          },
+        },
+        productGroup: null,
+        baseUnit: {
+          uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+          symbol: 'MT AIR',
+          title: 'Metric Ton Air',
+        },
+        unit2: null,
+        unit3: null,
+        limitNearlyFull: null,
+        limitNearlyEmpty: null,
+      },
+      {
+        classType: 'TANK',
+        uuid: '131f6e1b-4f27-4cae-8174-d5966f0f750c',
+        quay: {
+          uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+          id: 'ST300',
+          title: 'ST300',
+          companyId: 'ST300',
+        },
+        name: '1004',
+        tank: {
+          uuid: 'e9652e74-8de2-4d8f-9a17-456439a169fd',
+          code: '1004',
+          inactive: null,
+          inactivityType: null,
+          loadingPlaces: null,
+          loadingStation: null,
+          decisionMatrixProduct: null,
+          physicalTankQualityPercentage: null,
+          upgradePercentage: null,
+        },
+        product: {
+          uuid: '9785fd5d-6782-40e5-9af2-6b7b1e3dfe64',
+          id: 'P0000045',
+          quay: {
+            uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+            id: 'ST300',
+            title: 'ST300',
+            companyId: 'ST300',
+          },
+          status: 'ACTIVE',
+          overallApproval: 'MANUALLY_RELEASED',
+          client: {
+            id: 'K0050',
+            uuid: 'cc423e20-6533-4d6e-a058-ef3aa8cf0ea2',
+            name: 'TOTSA Total Oil Trading SA',
+          },
+          mainProduct: {
+            uuid: '09312045-f999-411a-9156-d0b9af8aef9b',
+            description: 'JET',
+            group: {
+              uuid: '7cb0785b-f02f-49fb-86a6-45dd836a6faa',
+              title: 'JET',
+            },
+          },
+          productProperty: {
+            uuid: 'a64f879a-53ee-4a5b-962d-dcb145b88dd4',
+            title: ' ',
+          },
+          quality: {
+            uuid: 'f4a1d27f-01da-44d3-a453-d39f80c6826e',
+            title: ' ',
+          },
+          manufacturer: {
+            uuid: '580e9c9f-e2a6-4c6f-834f-55c393ddb285',
+            code: 'ASTRO BIOPLANT',
+            title: 'Astro Bioplant',
+            address: null,
+            blockages: [],
+            types: ['PRODUCT_MANUFACTURER'],
+            vatNumber: null,
+            venNumber: null,
+          },
+          baseUnit: {
+            uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+            symbol: 'MT AIR',
+            title: 'Metric Ton Air',
+          },
+          unit2: {
+            uuid: 'daea0af5-de5d-46b3-a07e-507026137a15',
+            symbol: 'm3',
+            title: 'Cubic Meter',
+          },
+          unit3: {
+            uuid: '1a76317f-5304-4187-b336-ae3fd37186be',
+            symbol: 'MT VAC',
+            title: 'Metric Ton Vaccum',
+          },
+          type: 'COMMERCIAL_PRODUCT',
+          animalByProducts: null,
+          unNumber: null,
+        },
+        productGroup: null,
+        baseUnit: {
+          uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+          symbol: 'MT AIR',
+          title: 'Metric Ton Air',
+        },
+        unit2: null,
+        unit3: null,
+        limitNearlyFull: null,
+        limitNearlyEmpty: null,
+      },
+    ],
+    sofStockLocationNames: [],
+    wmsStatus: null,
+    invoicingCode: {
+      uuid: 'a491ac1e-c41c-44e8-9a99-bbe307031be1',
+      code: 'ULSD',
+    },
+    createdOn: '2023-08-29T09:43:17.5318208Z',
+  },
+  {
+    uuid: '142dc29e-595e-4d45-a048-4f56714ef322',
+    contractUuid: 'f3669b80-9237-4ab0-b107-13787bc894ab',
+    originalOrderUuid: null,
+    status: 'BOOKED',
+    expeditionLogStatus: 'ACTIVATED',
+    plannedDate: null,
+    clientReferences: {
+      reference1: '123',
+      reference2: '',
+      reference3: null,
+    },
+    transporter: {
+      title: 'TOTSA RELATION',
+      code: '123',
+      address: {
+        addressLine1: 'valkstraat 58',
+        addressLine2: null,
+        postalCode: '2860',
+        city: 'skw',
+        countryName: 'BELGIUM / BELGIQUE',
+      },
+    },
+    collector: null,
+    addressee: {
+      title: 'TOTSA RELATION',
+      code: '123',
+      address: {
+        addressLine1: 'valkstraat 58',
+        addressLine2: null,
+        postalCode: '2860',
+        city: 'skw',
+        countryName: 'BELGIUM / BELGIQUE',
+      },
+    },
+    activity: {
+      uuid: '2cd71785-3722-4c6f-963b-d42300a9e87e',
+      classType: null,
+      description: 'Ex tank in truck',
+    },
+    intermodal: null,
+    fullFreight: null,
+    nominatedQuantityBaseUnit: {
+      amount: 12,
+      unit: {
+        uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+        symbol: 'MT AIR',
+      },
+    },
+    processedQuantityBaseUnit: {
+      amount: 1,
+      unit: null,
+    },
+    ediOrder: null,
+    lcsOrderFields: {
+      uuid: 'cab2eb30-b095-439b-968f-0ee88ac34beb',
+      customsStatus: null,
+      lcsInternal: null,
+      lcsInstructionsWeighbridgeClerk: null,
+    },
+    remarks: {
+      changes: [],
+      customerService: 'LCS',
+      internal: null,
+    },
+    additionalInstructions: null,
+    plannedStockLocations: [
+      {
+        classType: 'TANK',
+        uuid: 'fd9e1bbf-74b3-4fd2-b3e5-7c378580bf46',
+        quay: {
+          uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+          id: 'ST300',
+          title: 'ST300',
+          companyId: 'ST300',
+        },
+        name: '1201',
+        tank: {
+          uuid: 'a0ad2afb-d67d-4e39-a8c9-1ad7990c82dc',
+          code: '1201',
+          inactive: null,
+          inactivityType: null,
+          loadingPlaces: 3,
+          loadingStation: null,
+          decisionMatrixProduct: {
+            id: 'P0000076',
+            uuid: '8f229886-7648-4b77-a3bd-672a97bca741',
+            type: 'COMMERCIAL_PRODUCT',
+            mainProductDescription: null,
+            productPropertyTitle: null,
+            qualityTitle: null,
+            manufacturerTitle: null,
+            productGroupTitle: null,
+            productGroupUuid: null,
+            quayId: null,
+            extranetTypeCode: null,
+            unNumberCode: null,
+          },
+          physicalTankQualityPercentage: 32,
+          upgradePercentage: null,
+        },
+        product: {
+          uuid: '8f229886-7648-4b77-a3bd-672a97bca741',
+          id: 'P0000076',
+          quay: {
+            uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+            id: 'ST300',
+            title: 'ST300',
+            companyId: 'ST300',
+          },
+          status: 'ACTIVE',
+          overallApproval: 'MANUALLY_RELEASED',
+          client: {
+            id: 'K0050',
+            uuid: 'cc423e20-6533-4d6e-a058-ef3aa8cf0ea2',
+            name: 'TOTSA Total Oil Trading SA',
+          },
+          mainProduct: {
+            uuid: 'e77f18fe-037d-46de-bd0a-f33db9c17ff4',
+            description: 'UAN',
+            group: {
+              uuid: '4f0e8834-4b5f-46d0-9c8c-711c786bc302',
+              title: 'UAN',
+            },
+          },
+          productProperty: {
+            uuid: 'a64f879a-53ee-4a5b-962d-dcb145b88dd4',
+            title: ' ',
+          },
+          quality: {
+            uuid: '9373ea89-4260-4062-8c77-d19b8dae8d73',
+            title: '32%',
+          },
+          manufacturer: {
+            uuid: '580e9c9f-e2a6-4c6f-834f-55c393ddb285',
+            code: 'ASTRO BIOPLANT',
+            title: 'Astro Bioplant',
+            address: null,
+            blockages: [],
+            types: ['PRODUCT_MANUFACTURER'],
+            vatNumber: null,
+            venNumber: null,
+          },
+          baseUnit: {
+            uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+            symbol: 'MT AIR',
+            title: 'Metric Ton Air',
+          },
+          unit2: null,
+          unit3: null,
+          type: 'COMMERCIAL_PRODUCT',
+          animalByProducts: null,
+          unNumber: null,
+        },
+        productGroup: null,
+        baseUnit: {
+          uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+          symbol: 'MT AIR',
+          title: 'Metric Ton Air',
+        },
+        unit2: null,
+        unit3: null,
+        limitNearlyFull: null,
+        limitNearlyEmpty: null,
+      },
+    ],
+    allPlannedStockLocations: [
+      {
+        classType: 'TANK',
+        uuid: '18ed4d34-f717-4381-aaa1-198f34474401',
+        quay: {
+          uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+          id: 'ST300',
+          title: 'ST300',
+          companyId: 'ST300',
+        },
+        name: '1002',
+        tank: {
+          uuid: '8ef9000c-064c-42bf-b003-e547915eb9c2',
+          code: '1002',
+          inactive: null,
+          inactivityType: null,
+          loadingPlaces: 2,
+          loadingStation: null,
+          decisionMatrixProduct: null,
+          physicalTankQualityPercentage: null,
+          upgradePercentage: null,
+        },
+        product: {
+          uuid: '5345881c-c769-4cb0-90a8-826f81ec2347',
+          id: 'P0000038',
+          quay: {
+            uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+            id: 'ST300',
+            title: 'ST300',
+            companyId: 'ST300',
+          },
+          status: 'ACTIVE',
+          overallApproval: 'MANUALLY_RELEASED',
+          client: {
+            id: 'K0050',
+            uuid: 'cc423e20-6533-4d6e-a058-ef3aa8cf0ea2',
+            name: 'TOTSA Total Oil Trading SA',
+          },
+          mainProduct: {
+            uuid: '109a4e59-2426-4585-9908-a11e2a6c6fe7',
+            description: 'C8',
+            group: {
+              uuid: '2210753a-631f-4746-9a7b-91673f4a088c',
+              title: 'GASOLINES',
+            },
+          },
+          productProperty: {
+            uuid: 'a64f879a-53ee-4a5b-962d-dcb145b88dd4',
+            title: ' ',
+          },
+          quality: {
+            uuid: 'f4a1d27f-01da-44d3-a453-d39f80c6826e',
+            title: ' ',
+          },
+          manufacturer: {
+            uuid: '7f26e016-6002-4918-afbc-888095b2163b',
+            code: 'BP',
+            title: 'BP',
+            address: {
+              uuid: null,
+              addressLine1: 'abc',
+              addressLine2: null,
+              postalCode: null,
+              city: null,
+              country: null,
+            },
+            blockages: [],
+            types: ['CONSIGNOR', 'PRODUCT_MANUFACTURER'],
+            vatNumber: null,
+            venNumber: null,
+          },
+          baseUnit: {
+            uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+            symbol: 'MT AIR',
+            title: 'Metric Ton Air',
+          },
+          unit2: {
+            uuid: 'daea0af5-de5d-46b3-a07e-507026137a15',
+            symbol: 'm3',
+            title: 'Cubic Meter',
+          },
+          unit3: {
+            uuid: '1a76317f-5304-4187-b336-ae3fd37186be',
+            symbol: 'MT VAC',
+            title: 'Metric Ton Vaccum',
+          },
+          type: 'COMMERCIAL_PRODUCT',
+          animalByProducts: null,
+          unNumber: {
+            uuid: '595947d9-9e5a-4fb7-b685-8cb11ad70971',
+            code: '456',
+            number: '4545',
+          },
+        },
+        productGroup: null,
+        baseUnit: {
+          uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+          symbol: 'MT AIR',
+          title: 'Metric Ton Air',
+        },
+        unit2: null,
+        unit3: null,
+        limitNearlyFull: null,
+        limitNearlyEmpty: null,
+      },
+      {
+        classType: 'TANK',
+        uuid: 'fd9e1bbf-74b3-4fd2-b3e5-7c378580bf46',
+        quay: {
+          uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+          id: 'ST300',
+          title: 'ST300',
+          companyId: 'ST300',
+        },
+        name: '1201',
+        tank: {
+          uuid: 'a0ad2afb-d67d-4e39-a8c9-1ad7990c82dc',
+          code: '1201',
+          inactive: null,
+          inactivityType: null,
+          loadingPlaces: 3,
+          loadingStation: null,
+          decisionMatrixProduct: {
+            id: 'P0000076',
+            uuid: '8f229886-7648-4b77-a3bd-672a97bca741',
+            type: 'COMMERCIAL_PRODUCT',
+            mainProductDescription: null,
+            productPropertyTitle: null,
+            qualityTitle: null,
+            manufacturerTitle: null,
+            productGroupTitle: null,
+            productGroupUuid: null,
+            quayId: null,
+            extranetTypeCode: null,
+            unNumberCode: null,
+          },
+          physicalTankQualityPercentage: 32,
+          upgradePercentage: null,
+        },
+        product: {
+          uuid: '8f229886-7648-4b77-a3bd-672a97bca741',
+          id: 'P0000076',
+          quay: {
+            uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+            id: 'ST300',
+            title: 'ST300',
+            companyId: 'ST300',
+          },
+          status: 'ACTIVE',
+          overallApproval: 'MANUALLY_RELEASED',
+          client: {
+            id: 'K0050',
+            uuid: 'cc423e20-6533-4d6e-a058-ef3aa8cf0ea2',
+            name: 'TOTSA Total Oil Trading SA',
+          },
+          mainProduct: {
+            uuid: 'e77f18fe-037d-46de-bd0a-f33db9c17ff4',
+            description: 'UAN',
+            group: {
+              uuid: '4f0e8834-4b5f-46d0-9c8c-711c786bc302',
+              title: 'UAN',
+            },
+          },
+          productProperty: {
+            uuid: 'a64f879a-53ee-4a5b-962d-dcb145b88dd4',
+            title: ' ',
+          },
+          quality: {
+            uuid: '9373ea89-4260-4062-8c77-d19b8dae8d73',
+            title: '32%',
+          },
+          manufacturer: {
+            uuid: '580e9c9f-e2a6-4c6f-834f-55c393ddb285',
+            code: 'ASTRO BIOPLANT',
+            title: 'Astro Bioplant',
+            address: null,
+            blockages: [],
+            types: ['PRODUCT_MANUFACTURER'],
+            vatNumber: null,
+            venNumber: null,
+          },
+          baseUnit: {
+            uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+            symbol: 'MT AIR',
+            title: 'Metric Ton Air',
+          },
+          unit2: null,
+          unit3: null,
+          type: 'COMMERCIAL_PRODUCT',
+          animalByProducts: null,
+          unNumber: null,
+        },
+        productGroup: null,
+        baseUnit: {
+          uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+          symbol: 'MT AIR',
+          title: 'Metric Ton Air',
+        },
+        unit2: null,
+        unit3: null,
+        limitNearlyFull: null,
+        limitNearlyEmpty: null,
+      },
+      {
+        classType: 'TANK',
+        uuid: '131f6e1b-4f27-4cae-8174-d5966f0f750c',
+        quay: {
+          uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+          id: 'ST300',
+          title: 'ST300',
+          companyId: 'ST300',
+        },
+        name: '1004',
+        tank: {
+          uuid: 'e9652e74-8de2-4d8f-9a17-456439a169fd',
+          code: '1004',
+          inactive: null,
+          inactivityType: null,
+          loadingPlaces: null,
+          loadingStation: null,
+          decisionMatrixProduct: null,
+          physicalTankQualityPercentage: null,
+          upgradePercentage: null,
+        },
+        product: {
+          uuid: '9785fd5d-6782-40e5-9af2-6b7b1e3dfe64',
+          id: 'P0000045',
+          quay: {
+            uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+            id: 'ST300',
+            title: 'ST300',
+            companyId: 'ST300',
+          },
+          status: 'ACTIVE',
+          overallApproval: 'MANUALLY_RELEASED',
+          client: {
+            id: 'K0050',
+            uuid: 'cc423e20-6533-4d6e-a058-ef3aa8cf0ea2',
+            name: 'TOTSA Total Oil Trading SA',
+          },
+          mainProduct: {
+            uuid: '09312045-f999-411a-9156-d0b9af8aef9b',
+            description: 'JET',
+            group: {
+              uuid: '7cb0785b-f02f-49fb-86a6-45dd836a6faa',
+              title: 'JET',
+            },
+          },
+          productProperty: {
+            uuid: 'a64f879a-53ee-4a5b-962d-dcb145b88dd4',
+            title: ' ',
+          },
+          quality: {
+            uuid: 'f4a1d27f-01da-44d3-a453-d39f80c6826e',
+            title: ' ',
+          },
+          manufacturer: {
+            uuid: '580e9c9f-e2a6-4c6f-834f-55c393ddb285',
+            code: 'ASTRO BIOPLANT',
+            title: 'Astro Bioplant',
+            address: null,
+            blockages: [],
+            types: ['PRODUCT_MANUFACTURER'],
+            vatNumber: null,
+            venNumber: null,
+          },
+          baseUnit: {
+            uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+            symbol: 'MT AIR',
+            title: 'Metric Ton Air',
+          },
+          unit2: {
+            uuid: 'daea0af5-de5d-46b3-a07e-507026137a15',
+            symbol: 'm3',
+            title: 'Cubic Meter',
+          },
+          unit3: {
+            uuid: '1a76317f-5304-4187-b336-ae3fd37186be',
+            symbol: 'MT VAC',
+            title: 'Metric Ton Vaccum',
+          },
+          type: 'COMMERCIAL_PRODUCT',
+          animalByProducts: null,
+          unNumber: null,
+        },
+        productGroup: null,
+        baseUnit: {
+          uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+          symbol: 'MT AIR',
+          title: 'Metric Ton Air',
+        },
+        unit2: null,
+        unit3: null,
+        limitNearlyFull: null,
+        limitNearlyEmpty: null,
+      },
+    ],
+    sofStockLocationNames: ['1002', '1004'],
+    wmsStatus: null,
+    invoicingCode: {
+      uuid: 'a7ca836f-8ed7-40ea-a62a-d97ecdc9ac29',
+      code: 'GASOLINES',
+    },
+    createdOn: '2023-05-02T13:10:57.8720856Z',
+  },
+  {
+    uuid: '62dd2a55-2052-4996-9ee0-790ff6d40ed9',
+    contractUuid: 'f3669b80-9237-4ab0-b107-13787bc894ab',
+    originalOrderUuid: null,
+    status: 'DRAFT',
+    expeditionLogStatus: 'ACTIVATED',
+    plannedDate: '2023-04-27',
+    clientReferences: {
+      reference1: 'Test12345',
+      reference2: null,
+      reference3: null,
+    },
+    transporter: null,
+    collector: null,
+    addressee: null,
+    activity: {
+      uuid: '2cd71785-3722-4c6f-963b-d42300a9e87e',
+      classType: null,
+      description: 'Ex tank in truck',
+    },
+    intermodal: false,
+    fullFreight: null,
+    nominatedQuantityBaseUnit: {
+      amount: 19.43,
+      unit: {
+        uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+        symbol: 'MT AIR',
+      },
+    },
+    processedQuantityBaseUnit: null,
+    ediOrder: null,
+    lcsOrderFields: {
+      uuid: 'dddf3a84-da8f-4da7-a8a6-f84839a09bba',
+      customsStatus: 'BONDED',
+      lcsInternal: null,
+      lcsInstructionsWeighbridgeClerk: null,
+    },
+    remarks: null,
+    additionalInstructions: null,
+    plannedStockLocations: [
+      {
+        classType: 'TANK',
+        uuid: 'cb591845-08e0-432c-9597-6a6734fd414c',
+        quay: {
+          uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+          id: 'ST300',
+          title: 'ST300',
+          companyId: 'ST300',
+        },
+        name: '1001',
+        tank: {
+          uuid: 'd5ef51f4-ca55-4c3d-85e1-4458d4f91a5b',
+          code: '1001',
+          inactive: false,
+          inactivityType: 'BOTH',
+          loadingPlaces: 1,
+          loadingStation: null,
+          decisionMatrixProduct: null,
+          physicalTankQualityPercentage: 0,
+          upgradePercentage: 0,
+        },
+        product: null,
+        productGroup: {
+          uuid: '9a43cd77-225c-4e5d-87e0-e74fed94c46a',
+          title: 'FUEL OILS',
+        },
+        baseUnit: {
+          uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+          symbol: 'MT AIR',
+          title: 'Metric Ton Air',
+        },
+        unit2: {
+          uuid: 'daea0af5-de5d-46b3-a07e-507026137a15',
+          symbol: 'm3',
+          title: 'Cubic Meter',
+        },
+        unit3: null,
+        limitNearlyFull: null,
+        limitNearlyEmpty: null,
+      },
+    ],
+    allPlannedStockLocations: [
+      {
+        classType: 'TANK',
+        uuid: '18ed4d34-f717-4381-aaa1-198f34474401',
+        quay: {
+          uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+          id: 'ST300',
+          title: 'ST300',
+          companyId: 'ST300',
+        },
+        name: '1002',
+        tank: {
+          uuid: '8ef9000c-064c-42bf-b003-e547915eb9c2',
+          code: '1002',
+          inactive: null,
+          inactivityType: null,
+          loadingPlaces: 2,
+          loadingStation: null,
+          decisionMatrixProduct: null,
+          physicalTankQualityPercentage: null,
+          upgradePercentage: null,
+        },
+        product: {
+          uuid: '5345881c-c769-4cb0-90a8-826f81ec2347',
+          id: 'P0000038',
+          quay: {
+            uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+            id: 'ST300',
+            title: 'ST300',
+            companyId: 'ST300',
+          },
+          status: 'ACTIVE',
+          overallApproval: 'MANUALLY_RELEASED',
+          client: {
+            id: 'K0050',
+            uuid: 'cc423e20-6533-4d6e-a058-ef3aa8cf0ea2',
+            name: 'TOTSA Total Oil Trading SA',
+          },
+          mainProduct: {
+            uuid: '109a4e59-2426-4585-9908-a11e2a6c6fe7',
+            description: 'C8',
+            group: {
+              uuid: '2210753a-631f-4746-9a7b-91673f4a088c',
+              title: 'GASOLINES',
+            },
+          },
+          productProperty: {
+            uuid: 'a64f879a-53ee-4a5b-962d-dcb145b88dd4',
+            title: ' ',
+          },
+          quality: {
+            uuid: 'f4a1d27f-01da-44d3-a453-d39f80c6826e',
+            title: ' ',
+          },
+          manufacturer: {
+            uuid: '7f26e016-6002-4918-afbc-888095b2163b',
+            code: 'BP',
+            title: 'BP',
+            address: {
+              uuid: null,
+              addressLine1: 'abc',
+              addressLine2: null,
+              postalCode: null,
+              city: null,
+              country: null,
+            },
+            blockages: [],
+            types: ['CONSIGNOR', 'PRODUCT_MANUFACTURER'],
+            vatNumber: null,
+            venNumber: null,
+          },
+          baseUnit: {
+            uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+            symbol: 'MT AIR',
+            title: 'Metric Ton Air',
+          },
+          unit2: {
+            uuid: 'daea0af5-de5d-46b3-a07e-507026137a15',
+            symbol: 'm3',
+            title: 'Cubic Meter',
+          },
+          unit3: {
+            uuid: '1a76317f-5304-4187-b336-ae3fd37186be',
+            symbol: 'MT VAC',
+            title: 'Metric Ton Vaccum',
+          },
+          type: 'COMMERCIAL_PRODUCT',
+          animalByProducts: null,
+          unNumber: {
+            uuid: '595947d9-9e5a-4fb7-b685-8cb11ad70971',
+            code: '456',
+            number: '4545',
+          },
+        },
+        productGroup: null,
+        baseUnit: {
+          uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+          symbol: 'MT AIR',
+          title: 'Metric Ton Air',
+        },
+        unit2: null,
+        unit3: null,
+        limitNearlyFull: null,
+        limitNearlyEmpty: null,
+      },
+      {
+        classType: 'TANK',
+        uuid: 'cb591845-08e0-432c-9597-6a6734fd414c',
+        quay: {
+          uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+          id: 'ST300',
+          title: 'ST300',
+          companyId: 'ST300',
+        },
+        name: '1001',
+        tank: {
+          uuid: 'd5ef51f4-ca55-4c3d-85e1-4458d4f91a5b',
+          code: '1001',
+          inactive: false,
+          inactivityType: 'BOTH',
+          loadingPlaces: 1,
+          loadingStation: null,
+          decisionMatrixProduct: null,
+          physicalTankQualityPercentage: 0,
+          upgradePercentage: 0,
+        },
+        product: null,
+        productGroup: {
+          uuid: '9a43cd77-225c-4e5d-87e0-e74fed94c46a',
+          title: 'FUEL OILS',
+        },
+        baseUnit: {
+          uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+          symbol: 'MT AIR',
+          title: 'Metric Ton Air',
+        },
+        unit2: {
+          uuid: 'daea0af5-de5d-46b3-a07e-507026137a15',
+          symbol: 'm3',
+          title: 'Cubic Meter',
+        },
+        unit3: null,
+        limitNearlyFull: null,
+        limitNearlyEmpty: null,
+      },
+      {
+        classType: 'TANK',
+        uuid: '131f6e1b-4f27-4cae-8174-d5966f0f750c',
+        quay: {
+          uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+          id: 'ST300',
+          title: 'ST300',
+          companyId: 'ST300',
+        },
+        name: '1004',
+        tank: {
+          uuid: 'e9652e74-8de2-4d8f-9a17-456439a169fd',
+          code: '1004',
+          inactive: null,
+          inactivityType: null,
+          loadingPlaces: null,
+          loadingStation: null,
+          decisionMatrixProduct: null,
+          physicalTankQualityPercentage: null,
+          upgradePercentage: null,
+        },
+        product: {
+          uuid: '9785fd5d-6782-40e5-9af2-6b7b1e3dfe64',
+          id: 'P0000045',
+          quay: {
+            uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+            id: 'ST300',
+            title: 'ST300',
+            companyId: 'ST300',
+          },
+          status: 'ACTIVE',
+          overallApproval: 'MANUALLY_RELEASED',
+          client: {
+            id: 'K0050',
+            uuid: 'cc423e20-6533-4d6e-a058-ef3aa8cf0ea2',
+            name: 'TOTSA Total Oil Trading SA',
+          },
+          mainProduct: {
+            uuid: '09312045-f999-411a-9156-d0b9af8aef9b',
+            description: 'JET',
+            group: {
+              uuid: '7cb0785b-f02f-49fb-86a6-45dd836a6faa',
+              title: 'JET',
+            },
+          },
+          productProperty: {
+            uuid: 'a64f879a-53ee-4a5b-962d-dcb145b88dd4',
+            title: ' ',
+          },
+          quality: {
+            uuid: 'f4a1d27f-01da-44d3-a453-d39f80c6826e',
+            title: ' ',
+          },
+          manufacturer: {
+            uuid: '580e9c9f-e2a6-4c6f-834f-55c393ddb285',
+            code: 'ASTRO BIOPLANT',
+            title: 'Astro Bioplant',
+            address: null,
+            blockages: [],
+            types: ['PRODUCT_MANUFACTURER'],
+            vatNumber: null,
+            venNumber: null,
+          },
+          baseUnit: {
+            uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+            symbol: 'MT AIR',
+            title: 'Metric Ton Air',
+          },
+          unit2: {
+            uuid: 'daea0af5-de5d-46b3-a07e-507026137a15',
+            symbol: 'm3',
+            title: 'Cubic Meter',
+          },
+          unit3: {
+            uuid: '1a76317f-5304-4187-b336-ae3fd37186be',
+            symbol: 'MT VAC',
+            title: 'Metric Ton Vaccum',
+          },
+          type: 'COMMERCIAL_PRODUCT',
+          animalByProducts: null,
+          unNumber: null,
+        },
+        productGroup: null,
+        baseUnit: {
+          uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+          symbol: 'MT AIR',
+          title: 'Metric Ton Air',
+        },
+        unit2: null,
+        unit3: null,
+        limitNearlyFull: null,
+        limitNearlyEmpty: null,
+      },
+    ],
+    sofStockLocationNames: [],
+    wmsStatus: null,
+    invoicingCode: {
+      uuid: 'a51e95b0-e0c2-4206-9116-eec6665468fb',
+      code: 'HS GO',
+    },
+    createdOn: '2023-04-04T11:52:34.6909913Z',
+  },
+  {
+    uuid: '6652e110-ec69-4013-ac7d-bd05ac76acac',
+    contractUuid: 'f3669b80-9237-4ab0-b107-13787bc894ab',
+    originalOrderUuid: null,
+    status: 'BOOKED',
+    expeditionLogStatus: 'ACTIVATED',
+    plannedDate: null,
+    clientReferences: {
+      reference1: '213',
+      reference2: null,
+      reference3: null,
+    },
+    transporter: {
+      title: 'TOTSA RELATION',
+      code: '123',
+      address: {
+        addressLine1: 'valkstraat 58',
+        addressLine2: null,
+        postalCode: '2860',
+        city: 'skw',
+        countryName: 'BELGIUM / BELGIQUE',
+      },
+    },
+    collector: null,
+    addressee: {
+      title: 'TOTSA RELATION',
+      code: '123',
+      address: {
+        addressLine1: 'valkstraat 58',
+        addressLine2: null,
+        postalCode: '2860',
+        city: 'skw',
+        countryName: 'BELGIUM / BELGIQUE',
+      },
+    },
+    activity: {
+      uuid: '2cd71785-3722-4c6f-963b-d42300a9e87e',
+      classType: null,
+      description: 'Ex tank in truck',
+    },
+    intermodal: true,
+    fullFreight: null,
+    nominatedQuantityBaseUnit: {
+      amount: 20,
+      unit: {
+        uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+        symbol: 'MT AIR',
+      },
+    },
+    processedQuantityBaseUnit: {
+      amount: 133,
+      unit: null,
+    },
+    ediOrder: null,
+    lcsOrderFields: {
+      uuid: 'c5f9b9e3-88ae-4511-9796-348f150f1922',
+      customsStatus: null,
+      lcsInternal: null,
+      lcsInstructionsWeighbridgeClerk: null,
+    },
+    remarks: null,
+    additionalInstructions: null,
+    plannedStockLocations: [],
+    allPlannedStockLocations: [
+      {
+        classType: 'TANK',
+        uuid: '18ed4d34-f717-4381-aaa1-198f34474401',
+        quay: {
+          uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+          id: 'ST300',
+          title: 'ST300',
+          companyId: 'ST300',
+        },
+        name: '1002',
+        tank: {
+          uuid: '8ef9000c-064c-42bf-b003-e547915eb9c2',
+          code: '1002',
+          inactive: null,
+          inactivityType: null,
+          loadingPlaces: 2,
+          loadingStation: null,
+          decisionMatrixProduct: null,
+          physicalTankQualityPercentage: null,
+          upgradePercentage: null,
+        },
+        product: {
+          uuid: '5345881c-c769-4cb0-90a8-826f81ec2347',
+          id: 'P0000038',
+          quay: {
+            uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+            id: 'ST300',
+            title: 'ST300',
+            companyId: 'ST300',
+          },
+          status: 'ACTIVE',
+          overallApproval: 'MANUALLY_RELEASED',
+          client: {
+            id: 'K0050',
+            uuid: 'cc423e20-6533-4d6e-a058-ef3aa8cf0ea2',
+            name: 'TOTSA Total Oil Trading SA',
+          },
+          mainProduct: {
+            uuid: '109a4e59-2426-4585-9908-a11e2a6c6fe7',
+            description: 'C8',
+            group: {
+              uuid: '2210753a-631f-4746-9a7b-91673f4a088c',
+              title: 'GASOLINES',
+            },
+          },
+          productProperty: {
+            uuid: 'a64f879a-53ee-4a5b-962d-dcb145b88dd4',
+            title: ' ',
+          },
+          quality: {
+            uuid: 'f4a1d27f-01da-44d3-a453-d39f80c6826e',
+            title: ' ',
+          },
+          manufacturer: {
+            uuid: '7f26e016-6002-4918-afbc-888095b2163b',
+            code: 'BP',
+            title: 'BP',
+            address: {
+              uuid: null,
+              addressLine1: 'abc',
+              addressLine2: null,
+              postalCode: null,
+              city: null,
+              country: null,
+            },
+            blockages: [],
+            types: ['CONSIGNOR', 'PRODUCT_MANUFACTURER'],
+            vatNumber: null,
+            venNumber: null,
+          },
+          baseUnit: {
+            uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+            symbol: 'MT AIR',
+            title: 'Metric Ton Air',
+          },
+          unit2: {
+            uuid: 'daea0af5-de5d-46b3-a07e-507026137a15',
+            symbol: 'm3',
+            title: 'Cubic Meter',
+          },
+          unit3: {
+            uuid: '1a76317f-5304-4187-b336-ae3fd37186be',
+            symbol: 'MT VAC',
+            title: 'Metric Ton Vaccum',
+          },
+          type: 'COMMERCIAL_PRODUCT',
+          animalByProducts: null,
+          unNumber: {
+            uuid: '595947d9-9e5a-4fb7-b685-8cb11ad70971',
+            code: '456',
+            number: '4545',
+          },
+        },
+        productGroup: null,
+        baseUnit: {
+          uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+          symbol: 'MT AIR',
+          title: 'Metric Ton Air',
+        },
+        unit2: null,
+        unit3: null,
+        limitNearlyFull: null,
+        limitNearlyEmpty: null,
+      },
+      {
+        classType: 'TANK',
+        uuid: '131f6e1b-4f27-4cae-8174-d5966f0f750c',
+        quay: {
+          uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+          id: 'ST300',
+          title: 'ST300',
+          companyId: 'ST300',
+        },
+        name: '1004',
+        tank: {
+          uuid: 'e9652e74-8de2-4d8f-9a17-456439a169fd',
+          code: '1004',
+          inactive: null,
+          inactivityType: null,
+          loadingPlaces: null,
+          loadingStation: null,
+          decisionMatrixProduct: null,
+          physicalTankQualityPercentage: null,
+          upgradePercentage: null,
+        },
+        product: {
+          uuid: '9785fd5d-6782-40e5-9af2-6b7b1e3dfe64',
+          id: 'P0000045',
+          quay: {
+            uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+            id: 'ST300',
+            title: 'ST300',
+            companyId: 'ST300',
+          },
+          status: 'ACTIVE',
+          overallApproval: 'MANUALLY_RELEASED',
+          client: {
+            id: 'K0050',
+            uuid: 'cc423e20-6533-4d6e-a058-ef3aa8cf0ea2',
+            name: 'TOTSA Total Oil Trading SA',
+          },
+          mainProduct: {
+            uuid: '09312045-f999-411a-9156-d0b9af8aef9b',
+            description: 'JET',
+            group: {
+              uuid: '7cb0785b-f02f-49fb-86a6-45dd836a6faa',
+              title: 'JET',
+            },
+          },
+          productProperty: {
+            uuid: 'a64f879a-53ee-4a5b-962d-dcb145b88dd4',
+            title: ' ',
+          },
+          quality: {
+            uuid: 'f4a1d27f-01da-44d3-a453-d39f80c6826e',
+            title: ' ',
+          },
+          manufacturer: {
+            uuid: '580e9c9f-e2a6-4c6f-834f-55c393ddb285',
+            code: 'ASTRO BIOPLANT',
+            title: 'Astro Bioplant',
+            address: null,
+            blockages: [],
+            types: ['PRODUCT_MANUFACTURER'],
+            vatNumber: null,
+            venNumber: null,
+          },
+          baseUnit: {
+            uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+            symbol: 'MT AIR',
+            title: 'Metric Ton Air',
+          },
+          unit2: {
+            uuid: 'daea0af5-de5d-46b3-a07e-507026137a15',
+            symbol: 'm3',
+            title: 'Cubic Meter',
+          },
+          unit3: {
+            uuid: '1a76317f-5304-4187-b336-ae3fd37186be',
+            symbol: 'MT VAC',
+            title: 'Metric Ton Vaccum',
+          },
+          type: 'COMMERCIAL_PRODUCT',
+          animalByProducts: null,
+          unNumber: null,
+        },
+        productGroup: null,
+        baseUnit: {
+          uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+          symbol: 'MT AIR',
+          title: 'Metric Ton Air',
+        },
+        unit2: null,
+        unit3: null,
+        limitNearlyFull: null,
+        limitNearlyEmpty: null,
+      },
+    ],
+    sofStockLocationNames: [],
+    wmsStatus: null,
+    invoicingCode: {
+      uuid: 'a7ca836f-8ed7-40ea-a62a-d97ecdc9ac29',
+      code: 'GASOLINES',
+    },
+    createdOn: '2023-03-29T19:21:09.7436712Z',
+  },
+  {
+    uuid: 'ca5f6fd6-bdfe-4388-bcc5-94e470de5eba',
+    contractUuid: 'f3669b80-9237-4ab0-b107-13787bc894ab',
+    originalOrderUuid: null,
+    status: 'BOOKED',
+    expeditionLogStatus: 'ACTIVATED',
+    plannedDate: '2023-03-23',
+    clientReferences: {
+      reference1: 'Blend test 2/1',
+      reference2: null,
+      reference3: null,
+    },
+    transporter: {
+      title: 'TOTSA RELATION',
+      code: '123',
+      address: {
+        addressLine1: 'valkstraat 58',
+        addressLine2: null,
+        postalCode: '2860',
+        city: 'skw',
+        countryName: 'BELGIUM / BELGIQUE',
+      },
+    },
+    collector: null,
+    addressee: {
+      title: 'TOTSA RELATION',
+      code: '123',
+      address: {
+        addressLine1: 'valkstraat 58',
+        addressLine2: null,
+        postalCode: '2860',
+        city: 'skw',
+        countryName: 'BELGIUM / BELGIQUE',
+      },
+    },
+    activity: {
+      uuid: '2cd71785-3722-4c6f-963b-d42300a9e87e',
+      classType: null,
+      description: 'Ex tank in truck',
+    },
+    intermodal: true,
+    fullFreight: null,
+    nominatedQuantityBaseUnit: {
+      amount: 20,
+      unit: {
+        uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+        symbol: 'MT AIR',
+      },
+    },
+    processedQuantityBaseUnit: {
+      amount: -20,
+      unit: null,
+    },
+    ediOrder: null,
+    lcsOrderFields: {
+      uuid: '7bcbdf04-8a5d-4371-9b1e-dcbc691b493a',
+      customsStatus: null,
+      lcsInternal: null,
+      lcsInstructionsWeighbridgeClerk: null,
+    },
+    remarks: null,
+    additionalInstructions: null,
+    plannedStockLocations: [],
+    allPlannedStockLocations: [
+      {
+        classType: 'TANK',
+        uuid: '18ed4d34-f717-4381-aaa1-198f34474401',
+        quay: {
+          uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+          id: 'ST300',
+          title: 'ST300',
+          companyId: 'ST300',
+        },
+        name: '1002',
+        tank: {
+          uuid: '8ef9000c-064c-42bf-b003-e547915eb9c2',
+          code: '1002',
+          inactive: null,
+          inactivityType: null,
+          loadingPlaces: 2,
+          loadingStation: null,
+          decisionMatrixProduct: null,
+          physicalTankQualityPercentage: null,
+          upgradePercentage: null,
+        },
+        product: {
+          uuid: '5345881c-c769-4cb0-90a8-826f81ec2347',
+          id: 'P0000038',
+          quay: {
+            uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+            id: 'ST300',
+            title: 'ST300',
+            companyId: 'ST300',
+          },
+          status: 'ACTIVE',
+          overallApproval: 'MANUALLY_RELEASED',
+          client: {
+            id: 'K0050',
+            uuid: 'cc423e20-6533-4d6e-a058-ef3aa8cf0ea2',
+            name: 'TOTSA Total Oil Trading SA',
+          },
+          mainProduct: {
+            uuid: '109a4e59-2426-4585-9908-a11e2a6c6fe7',
+            description: 'C8',
+            group: {
+              uuid: '2210753a-631f-4746-9a7b-91673f4a088c',
+              title: 'GASOLINES',
+            },
+          },
+          productProperty: {
+            uuid: 'a64f879a-53ee-4a5b-962d-dcb145b88dd4',
+            title: ' ',
+          },
+          quality: {
+            uuid: 'f4a1d27f-01da-44d3-a453-d39f80c6826e',
+            title: ' ',
+          },
+          manufacturer: {
+            uuid: '7f26e016-6002-4918-afbc-888095b2163b',
+            code: 'BP',
+            title: 'BP',
+            address: {
+              uuid: null,
+              addressLine1: 'abc',
+              addressLine2: null,
+              postalCode: null,
+              city: null,
+              country: null,
+            },
+            blockages: [],
+            types: ['CONSIGNOR', 'PRODUCT_MANUFACTURER'],
+            vatNumber: null,
+            venNumber: null,
+          },
+          baseUnit: {
+            uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+            symbol: 'MT AIR',
+            title: 'Metric Ton Air',
+          },
+          unit2: {
+            uuid: 'daea0af5-de5d-46b3-a07e-507026137a15',
+            symbol: 'm3',
+            title: 'Cubic Meter',
+          },
+          unit3: {
+            uuid: '1a76317f-5304-4187-b336-ae3fd37186be',
+            symbol: 'MT VAC',
+            title: 'Metric Ton Vaccum',
+          },
+          type: 'COMMERCIAL_PRODUCT',
+          animalByProducts: null,
+          unNumber: {
+            uuid: '595947d9-9e5a-4fb7-b685-8cb11ad70971',
+            code: '456',
+            number: '4545',
+          },
+        },
+        productGroup: null,
+        baseUnit: {
+          uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+          symbol: 'MT AIR',
+          title: 'Metric Ton Air',
+        },
+        unit2: null,
+        unit3: null,
+        limitNearlyFull: null,
+        limitNearlyEmpty: null,
+      },
+      {
+        classType: 'TANK',
+        uuid: '131f6e1b-4f27-4cae-8174-d5966f0f750c',
+        quay: {
+          uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+          id: 'ST300',
+          title: 'ST300',
+          companyId: 'ST300',
+        },
+        name: '1004',
+        tank: {
+          uuid: 'e9652e74-8de2-4d8f-9a17-456439a169fd',
+          code: '1004',
+          inactive: null,
+          inactivityType: null,
+          loadingPlaces: null,
+          loadingStation: null,
+          decisionMatrixProduct: null,
+          physicalTankQualityPercentage: null,
+          upgradePercentage: null,
+        },
+        product: {
+          uuid: '9785fd5d-6782-40e5-9af2-6b7b1e3dfe64',
+          id: 'P0000045',
+          quay: {
+            uuid: '50068068-b201-4402-8d8e-1132a40dc8d6',
+            id: 'ST300',
+            title: 'ST300',
+            companyId: 'ST300',
+          },
+          status: 'ACTIVE',
+          overallApproval: 'MANUALLY_RELEASED',
+          client: {
+            id: 'K0050',
+            uuid: 'cc423e20-6533-4d6e-a058-ef3aa8cf0ea2',
+            name: 'TOTSA Total Oil Trading SA',
+          },
+          mainProduct: {
+            uuid: '09312045-f999-411a-9156-d0b9af8aef9b',
+            description: 'JET',
+            group: {
+              uuid: '7cb0785b-f02f-49fb-86a6-45dd836a6faa',
+              title: 'JET',
+            },
+          },
+          productProperty: {
+            uuid: 'a64f879a-53ee-4a5b-962d-dcb145b88dd4',
+            title: ' ',
+          },
+          quality: {
+            uuid: 'f4a1d27f-01da-44d3-a453-d39f80c6826e',
+            title: ' ',
+          },
+          manufacturer: {
+            uuid: '580e9c9f-e2a6-4c6f-834f-55c393ddb285',
+            code: 'ASTRO BIOPLANT',
+            title: 'Astro Bioplant',
+            address: null,
+            blockages: [],
+            types: ['PRODUCT_MANUFACTURER'],
+            vatNumber: null,
+            venNumber: null,
+          },
+          baseUnit: {
+            uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+            symbol: 'MT AIR',
+            title: 'Metric Ton Air',
+          },
+          unit2: {
+            uuid: 'daea0af5-de5d-46b3-a07e-507026137a15',
+            symbol: 'm3',
+            title: 'Cubic Meter',
+          },
+          unit3: {
+            uuid: '1a76317f-5304-4187-b336-ae3fd37186be',
+            symbol: 'MT VAC',
+            title: 'Metric Ton Vaccum',
+          },
+          type: 'COMMERCIAL_PRODUCT',
+          animalByProducts: null,
+          unNumber: null,
+        },
+        productGroup: null,
+        baseUnit: {
+          uuid: '4abc4e87-252a-4be1-9d20-6ab0f196cdf7',
+          symbol: 'MT AIR',
+          title: 'Metric Ton Air',
+        },
+        unit2: null,
+        unit3: null,
+        limitNearlyFull: null,
+        limitNearlyEmpty: null,
+      },
+    ],
+    sofStockLocationNames: [],
+    wmsStatus: null,
+    invoicingCode: {
+      uuid: 'a7ca836f-8ed7-40ea-a62a-d97ecdc9ac29',
+      code: 'GASOLINES',
+    },
+    createdOn: '2023-03-23T15:19:09.3232871Z',
+  },
+];
 
 @Component({
   selector: 'lab900-table-example',
+  standalone: true,
+  imports: [
+    Lab900TableComponent,
+    Lab900TableTopContentDirective,
+    Lab900TableHeaderContentDirective,
+    Lab900TableEmptyDirective,
+  ],
   template: ` <lab900-table
-    [pageSizeConfig]="{ hidePageSize: true, pageSizeOptions: [5, 10] }"
     [tableCells]="tableCells"
     [sort]="sort"
     (sortChange)="sortChange($event)"
     [data]="mockData"
-    [paging]="paging"
-    [tableActionsBack]="tableActions"
     [tableHeaderActions]="tableHeaderActions"
-    [toggleColumns]="false"
     [toggleAndMoveColumns]="true"
     filterIcon="settings"
-    [selectableRows]="true"
-    [selectableRowsOptions]="{ selectedItems: selectedItems, disabled: true }"
-    [onRowClick]="rowClick"
     [multiSort]="true"
     [rowClass]="getRowClass"
     (tableCellsFiltered)="filtered($event)"
     [maxColumnWidth]="'200px'"
-    [tableFooterActions]="tableFooterActions"
     [preFooterTitle]="'Quantity Total'"
+    (cellValueChanged)="cellValueChanged($event)"
+    [trackByTableFn]="trackByTableFn"
+    [onRowClick]="onRowClick"
   >
     <div *lab900TableTopContent>Custom top content</div>
     <div *lab900TableHeaderContent>Custom header</div>
-    <div *lab900TableCustomCell="let data">
-      <div *ngIf="data.cell.key === 'active'">
-        <mat-checkbox color="primary" [checked]="data.element?.active"></mat-checkbox>
-      </div>
-    </div>
+
     <div *lab900TableEmpty>
       <div class="no-results">
         <p>No results template (can be anything)</p>
@@ -42,17 +2851,6 @@ import { ActionButton, Lab900Sort, Paging, TableCell } from '@lab900/ui';
 })
 export class TableExampleComponent {
   public sort: Lab900Sort[] = [{ id: 'id', direction: 'asc' }];
-
-  public tableFooterActions: ActionButton[] = [
-    {
-      label: 'Kies een locatie',
-      type: 'flat',
-    },
-    {
-      label: 'Button',
-      type: 'stroked',
-    },
-  ];
 
   public tableHeaderActions: ActionButton[] = [
     {
@@ -76,10 +2874,12 @@ export class TableExampleComponent {
         {
           label: 'Word',
           type: 'stroked',
+          action: () => console.log('Word'),
         },
         {
           label: 'PDF',
           type: 'stroked',
+          action: () => console.log('PDF'),
         },
       ],
     },
@@ -117,119 +2917,88 @@ export class TableExampleComponent {
     },
   ];
 
-  public mockData: any[] = [
-    {
-      name: 'John Cena',
-      nameLong: 'John Cena, the guy from wrestling and movies and stuff, and also just from the memes',
-      id: 1,
-      nested: {
-        test: 'xxx',
-      },
-      email: 'john@cena.com',
-      city: 'New York City',
-      quantity: 123,
-    },
-    {
-      name: 'A name',
-      nameLong: 'A very very veeeeeeeeryyy loooooooooong name',
-      id: 2,
-      active: true,
-      nested: {},
-      email: 'b@name.com',
-      quantity: 456,
-    },
-    {
-      name: '',
-      nameLong: 'A name very long :)',
-      id: 3,
-      quantity: 789,
-    },
-  ];
+  public mockData: any[] = mockData;
 
   public selectedItems = [this.mockData[0]];
 
-  public paging: Paging = {
-    pageIndex: 0,
-    pageSize: 5,
-    totalItems: this.mockData.length,
-  };
-
   public tableCells: TableCell[] = [
     {
-      key: 'name',
-      label: 'Name',
-      cellHeaderIcon: 'accessibility',
-      cellHeaderClass: 'center-cell',
-      sortable: true,
-      cellClass: 'clickable-cell',
-      cellTooltip: {
-        text: (data) => data.email,
-        tooltipOptions: { tooltipPosition: 'left' },
-      },
-      columnOrder: 0,
-      footer: '<a href="javascript:void(0)">Click here!</a>',
-      sticky: true,
+      key: 'expeditionLogStatus',
+      label: 'Status',
     },
     {
-      key: 'nameLong',
-      label: 'Long name',
+      key: 'clientReferences.reference1',
+      label: 'clientReferences.reference1',
       sortable: true,
-      cellClass: 'clickable-cell',
-      cellTooltip: { text: (data) => data.nameLong, onlyOnOverflow: true },
-      cellMaxWidth: '300px', // overrides the maxColumnWidth on the table
-      columnOrder: 1,
-      click: (data, cell, mouseEvent) => {
-        mouseEvent.stopPropagation();
-        console.log('click cell content > function params:', { data }, { cell }, { mouseEvent });
+      cellEditor: CellInputEditorComponent,
+      cellEditorOptions: <CellInputEditorOptions>{
+        placeholder: 'reference1',
+        valueChanged: ({ value, cell, row }: CellValueChangeEvent) => {
+          row[cell.key] = value;
+        },
       },
     },
     {
-      key: 'quantity',
-      label: 'Quantity',
-      cellClass: 'clickable-cell',
-      footer: (tableData) => {
-        return tableData.map((data) => data.quantity).reduce((valA, valB) => valA + valB, 0);
-      },
-      footerCellClass: 'table-footer-highlight',
-    },
-    {
-      key: 'id',
-      label: 'ID',
+      key: 'clientReferences.reference2',
+      label: 'clientReferences.reference2',
       sortable: true,
-      cellClass: 'clickable-cell',
-      width: '*',
-      columnOrder: 3,
+      cellEditor: CellInputEditorComponent,
+      cellEditorOptions: <CellInputEditorOptions>{
+        placeholder: 'reference2',
+
+        valueChanged: ({ value, cell, row }: CellValueChangeEvent) => {
+          row[cell.key] = value;
+        },
+      },
     },
     {
-      key: 'active',
-      label: 'Active',
-      customCellContent: true,
-      cellClass: 'center-cell',
-      columnOrder: 2,
+      key: 'clientReferences.reference3',
+      label: 'clientReferences.reference3',
+      sortable: true,
+      cellEditor: CellInputEditorComponent,
+      cellEditorOptions: <CellInputEditorOptions>{
+        placeholder: 'reference3',
+
+        valueChanged: ({ value, cell, row }: CellValueChangeEvent) => {
+          row[cell.key] = value;
+        },
+      },
     },
     {
-      key: 'email',
-      label: 'Email',
-      columnOrder: 4,
-      hide: true,
-    },
-    {
-      key: 'city',
-      label: 'City',
-      columnOrder: 5,
-      hide: true,
+      key: 'type',
+      label: 'NORMAL SELECT',
+      cellClass: (data) => (data.required ? 'table-cell-required-field' : ''),
+      width: '100px',
+      cellEditor: CellSelectEditorComponent,
+      cellRenderer: CellWithClearingRendererComponent,
+      cellFormatter: ({ type }) => type?.name ?? '',
+      cellEditorOptions: <CellSelectEditorOptions>{
+        panelWidth: '200px',
+        placeholder: 'Select a type 2',
+        compareWithFn: (a, b) => a?.id === b?.id,
+        optionLabelFn: (option) => option?.name ?? '?',
+        options: [
+          { id: 1, name: 'Type 1' },
+          { id: 2, name: 'Type 2' },
+        ],
+        valueChanged: ({ value, cell, row }: CellValueChangeEvent) => {
+          row[cell.key] = value;
+        },
+        disabled: (e) =>
+          (e as any).expeditionLogStatus === 'ACTIVATED' || false,
+      },
     },
   ];
+  public trackByTableFn: TrackByFunction<any> = (index, item) => item.id;
 
   public sortChange(sort: Lab900Sort[]): void {
     sort.forEach((s) => {
-      this.mockData.sort((a: any, b: any) => (a[s.id] < b[s.id] ? -1 : 1) * (s.direction === 'asc' ? 1 : -1));
+      this.mockData.sort(
+        (a: any, b: any) =>
+          (a[s.id] < b[s.id] ? -1 : 1) * (s.direction === 'asc' ? 1 : -1),
+      );
       this.mockData = [...this.mockData];
     });
-  }
-
-  public rowClick(row, i, event): void {
-    console.log('click row > function params:', { row }, { i }, { event });
   }
 
   public filtered(tableCells: TableCell[]): void {
@@ -238,5 +3007,17 @@ export class TableExampleComponent {
 
   public getRowClass(row: any): string {
     return row.active ? 'bg-green' : '';
+  }
+
+  public cellValueChanged(event: CellValueChangeEvent): void {
+    event.row[event.cell.key] = event.value;
+    setTimeout(() => {
+      this.mockData = [...this.mockData.map((d) => ({ ...d }))];
+    }, 200);
+    console.log('CellValueChangeEvent', event);
+  }
+
+  public onRowClick(row: any): void {
+    console.log('Row clicked', row);
   }
 }
